Calculate OH reactivity

Calculate OH reactivity of VOC time series in 25 degree celsius. Note: for Chinese VOC name, please also use English punctuation.

Usage

loh(
  df,
  unit = "ppbv",
  t = 25,
  p = 101.325,
  stcd = FALSE,
  sortd = TRUE,
  atk = TRUE,
  chn = FALSE,
  bvoc = TRUE
)

Arguments

df

dataframe contains time series.

unit

unit for VOC concentration. A character vector from these options: "ugm" or "ppbv". "ugm" means ug/m3. "ppbv" means part per billion volumn.

t

Temperature, in Degrees Celsius, used to convert data in micrograms per cubic meter to standard conditions (25 Degrees Celsius, 101.325 kPa). By default, t equals to 25 Degrees Celsius.

p

Pressure, in kPa, used to convert data in micrograms per cubic meter to standard conditions (25 Degrees Celsius, 101.325 kPa). By default, p equals to 101.325 kPa.

stcd

logical. Does it output the concentration in standard condition? The default vaule is FALSE.

sortd

logical value. It determines whether the VOC species are sorted or not. By default, sortd has value "TRUE". If TRUE, VOC species in time series will be arranged according to VOC group, relative molecular weight, and OH Rate Constant.

atk

logical. use kOH value from atk or not? If not, kOH comes from 'AopWin v1.92' will be used. The default vaule is TRUE.

chn

logical. Dose colnames present as Chinese? The default vaule is FALSE.

bvoc

logical. Whether you want to list BVOC as a separate VOC group? The default vaule is TRUE.

Value

a list contains 5 tables: KOH_Result: the matched KOH value results; LOH_Result: the LOH time series of VOC by species; LOH_Result_stat: the statistics of LOH of VOC by species; LOH_Result_group: the LOH time series of VOC classified by groups; LOH_Result_group_mean: the statistics of LOH of VOC according to major groups.

Details

The CAS number is matched for each VOC speices (from column name), and the OH Rate Constant is matched through the CAS number and used for time series calculation.
The OH Rate Constant comes from 'AopWin v1.92' in 25 degree celsius.

Examples

voc_loh=loh(voc)
summary(voc_loh)
#>                       Length Class      Mode
#> KOH_Result            8      data.frame list
#> LOH_Result            6      data.frame list
#> LOH_Result_stat       9      data.frame list
#> LOH_Result_group      4      data.frame list
#> LOH_Result_group_stat 9      data.frame list