Calculate ozone formation potential

Calculate Ozone Formation Potential (OFP) of VOC time series. Note: for Chinese VOC name, please also use English punctuation.

Usage

ofp(
  df,
  inunit = "ppbv",
  outunit = "ppbv",
  t = 25,
  p = 101.325,
  stcd = FALSE,
  sortd = TRUE,
  chn = FALSE,
  mtype = "usa",
  bvoc = TRUE
)

Arguments

df

dataframe contains time series.

inunit

input's unit for VOC concentration. A character vector from these options: "ugm" or "ppbv". "ugm" means ug/m3. "ppbv" means part per billion volumn. The default vaule is "ppbv".

outunit

output's unit for VOC concentration. A character from these options: "ugm" or "ppbv". "ugm" means ug/m3. "ppbv" means part per billion volumn. The default vaule is "ppbv".

t

Temperature, in Degrees Celsius, used to convert data in micrograms per cubic meter to standard conditions (25 Degrees Celsius, 101.325 kPa). By default, t equals to 25 Degrees Celsius.

p

Pressure, in kPa, used to convert data in micrograms per cubic meter to standard conditions (25 Degrees Celsius, 101.325 kPa). By default, p equals to 101.325 kPa.

stcd

logical. Does it output results in standard conditions? The default vaule is FALSE.

sortd

logical value. It determines whether the VOC species are sorted or not. By default, sortd has value "TRUE". If TRUE, VOC species in time series will be arranged according to VOC group, relative molecular weight, and MIR value.

chn

logical. Dose colnames present as Chinese? The default vaule is FALSE.

mtype

text. "usa" for MIR value from USA, "chn" for MIR value from CHINA.

bvoc

logical. Whether you want to list BVOC as a separate VOC group? The default vaule is TRUE.

Value

a list contains 5 tables: MIR_Result: the matched MIR value results; OFP_Result: the OFP time series of VOC by species; OFP_Result_stat: the statistics of OFP of VOC by species; OFP_Result_group: OFP the time series of VOC classified by groups; OFP_Result_group_stat: the statistics of OFP of VOC according to major groups.

Details

The CAS number is matched for each VOC speices (from column name), and the Maximum Incremental Reactivity (MIR) value is matched through the CAS number and used for time series calculation.
The MIR value comes from <https://ww2.arb.ca.gov/sites/default/files/classic/regact/2009/mir2009/mir10.pdf>, Zhang et al.(2021) <doi:10.5194/acp-21-11053-2021>.

Examples

ofp(voc)
#> $MIR_Result
#>             Name      CAS   Matched_Name   MIR MIR_type    MW   Group Raw_order
#> 3       n-Butane 106-97-8       n-butane  1.15      USA 58.12 Alkanes         3
#> 5    Cyclohexane 110-82-7    cyclohexane  1.25      USA 84.16 Alkanes         5
#> 1      Propylene 115-07-1        propene 11.66      USA 42.08 Alkenes         1
#> 4 trans-2-Butene 624-64-6 trans-2-butene 15.16      USA 56.11 Alkenes         4
#> 2      Acetylene  74-86-2      acetylene  0.95      USA 26.04 Alkynes         2
#> 
#> $OFP_Result
#>                    Time n.Butane Cyclohexane Propylene trans.2.Butene Acetylene
#> 1   2020-05-01 00:00:00    0.757       0.224     2.382          0.354     0.090
#> 2   2020-05-01 01:00:00    0.980       0.229     3.843          0.496     0.104
#> 3   2020-05-01 02:00:00    1.203       0.235     5.305          0.638     0.119
#> 4   2020-05-01 03:00:00    1.649       0.245     8.229          0.922     0.147
#> 5   2020-05-01 04:00:00    1.816       0.270     6.726          1.329     0.150
#> 6   2020-05-01 05:00:00    1.259       0.243     5.499          0.868     0.191
#> 7   2020-05-01 06:00:00    1.167       0.281     5.448          0.727     0.205
#> 8   2020-05-01 07:00:00    1.097       0.307     5.867          0.922     0.242
#> 9   2020-05-01 08:00:00    1.044       0.261     6.368          0.603     0.284
#> 10  2020-05-01 09:00:00    1.317       0.215     4.569          0.922     0.267
#> 11  2020-05-01 10:00:00    0.952       0.180     3.373          0.638     0.115
#> 12  2020-05-01 11:00:00    0.744       0.175     4.171          0.762     0.058
#> 13  2020-05-01 12:00:00    0.689       0.171     3.445          0.744     0.077
#> 14  2020-05-01 13:00:00    0.808       0.191     3.424          0.886     0.087
#> 15  2020-05-01 14:00:00    0.656       0.184     3.343          0.673     0.088
#> 16  2020-05-01 15:00:00    0.682       0.184     4.324          0.727     0.149
#> 17  2020-05-01 16:00:00    0.920       0.204     4.774          0.727     0.146
#> 18  2020-05-01 17:00:00    0.866       0.191     4.007          0.532     0.139
#> 19  2020-05-01 18:00:00    0.714       0.184     5.939          0.691     0.139
#> 20  2020-05-01 19:00:00    1.418       0.202     8.096          0.797     0.205
#> 21  2020-05-01 20:00:00    1.714       0.215    11.285          0.815     0.311
#> 22  2020-05-01 21:00:00    1.640       0.208    10.447          1.294     0.294
#> 23  2020-05-01 22:00:00    1.764       0.213     7.820          1.028     0.188
#> 24  2020-05-01 23:00:00    1.324       0.226     6.644          0.744     0.296
#> 25  2020-05-02 00:00:00    0.812       0.228     5.458          0.656     0.176
#> 26  2020-05-02 01:00:00    0.577       0.301     7.035          0.611     0.130
#> 27  2020-05-02 02:00:00    0.695       0.300     5.959          0.673     0.109
#> 28  2020-05-02 03:00:00    0.930       0.298     3.808          0.797     0.066
#> 29  2020-05-02 04:00:00    1.049       0.313     4.487          0.744     0.105
#> 30  2020-05-02 05:00:00    1.285       0.344     5.847          0.638     0.183
#> 31  2020-05-02 06:00:00    1.241       0.298     6.879          1.063     0.256
#> 32  2020-05-02 07:00:00    1.578       0.259     5.596          1.444     0.499
#> 33  2020-05-02 08:00:00    1.500       0.283     6.286          1.187     0.454
#> 34  2020-05-02 09:00:00    1.342       0.331     7.666          0.673     0.362
#> 35  2020-05-02 10:00:00    1.353       0.333     7.431          0.709     0.331
#> 36  2020-05-02 11:00:00    1.834       0.366     8.484          0.673     0.303
#> 37  2020-05-02 12:00:00    1.636       0.259     7.022          0.815     0.338
#> 38  2020-05-02 13:00:00    0.831       0.197     4.201          0.709     0.183
#> 39  2020-05-02 14:00:00    0.546       0.162     3.322          0.762     0.096
#> 40  2020-05-02 15:00:00    0.457       0.160     2.913          0.797     0.088
#> 41  2020-05-02 16:00:00    0.475       0.171     2.923          0.585     0.097
#> 42  2020-05-02 17:00:00    0.841       0.164     3.486          0.780     0.113
#> 43  2020-05-02 18:00:00    0.991       0.175     4.293          0.797     0.166
#> 44  2020-05-02 19:00:00    1.470       0.193     6.450          0.992     0.154
#> 45  2020-05-02 20:00:00    2.499       0.213    15.415          1.205     0.348
#> 46  2020-05-02 21:00:00    2.334       0.206    16.038          1.737     0.270
#> 47  2020-05-02 22:00:00    2.206       0.228    11.592          0.939     0.216
#> 48  2020-05-02 23:00:00    3.999       0.253    15.882          2.556     0.179
#> 49  2020-05-03 00:00:00    3.931       0.252    16.442          2.454     0.183
#> 50  2020-05-03 01:00:00    3.794       0.250    17.561          2.251     0.191
#> 51  2020-05-03 02:00:00    3.522       0.245    19.800          1.843     0.207
#> 52  2020-05-03 03:00:00    7.833       0.267    41.531          3.562     0.249
#> 53  2020-05-03 04:00:00    4.039       0.243    20.566          1.843     0.225
#> 54  2020-05-03 05:00:00    3.547       0.265    22.805          1.506     0.290
#> 55  2020-05-03 06:00:00    2.999       0.309    23.091          2.481     0.490
#> 56  2020-05-03 07:00:00    2.069       0.243    11.939          1.329     0.246
#> 57  2020-05-03 08:00:00    1.159       0.195     4.467          0.673     0.160
#> 58  2020-05-03 09:00:00    1.273       0.184     4.671          0.833     0.191
#> 59  2020-05-03 10:00:00    0.746       0.175     4.140          0.656     0.170
#> 60  2020-05-03 11:00:00    0.778       0.171     3.159          0.886     0.156
#> 61  2020-05-03 12:00:00    0.692       0.167     3.813          0.886     0.149
#> 62  2020-05-03 13:00:00    0.625       0.164     2.801          0.656     0.165
#> 63  2020-05-03 14:00:00    0.546       0.164     2.903          0.833     0.175
#> 64  2020-05-03 15:00:00    0.654       0.145     3.015          0.620     0.159
#> 65  2020-05-03 16:00:00    1.051       0.167     3.649          0.744     0.186
#> 66  2020-05-03 17:00:00    0.756       0.173     4.181          0.673     0.228
#> 67  2020-05-03 18:00:00    1.184       0.167     4.814          0.691     0.236
#> 68  2020-05-03 19:00:00    0.847       0.169     5.336          0.833     0.257
#> 69  2020-05-03 20:00:00    0.943       0.186     5.847          0.673     0.244
#> 70  2020-05-03 21:00:00    0.770       0.169     5.019          0.992     0.223
#> 71  2020-05-03 22:00:00    0.848       0.167     3.721          0.567     0.206
#> 72  2020-05-03 23:00:00    0.698       0.160     3.486          0.673     0.202
#> 73  2020-05-04 00:00:00    0.714       0.161     3.455          0.913     0.229
#> 74  2020-05-04 01:00:00    0.723       0.164     3.322          0.851     0.223
#> 75  2020-05-04 02:00:00    0.739       0.171     3.056          0.727     0.211
#> 76  2020-05-04 03:00:00    0.780       0.178     3.445          0.673     0.233
#> 77  2020-05-04 04:00:00    0.625       0.167     3.107          0.532     0.215
#> 78  2020-05-04 05:00:00    0.859       0.184     3.629          0.762     0.357
#> 79  2020-05-04 06:00:00    0.977       0.208     4.079          1.046     0.329
#> 80  2020-05-04 07:00:00    1.147       0.243     4.661          0.904     0.255
#> 81  2020-05-04 08:00:00    1.227       0.232     4.886          0.673     0.251
#> 82  2020-05-04 09:00:00    1.423       0.206     5.213          1.187     0.250
#> 83  2020-05-04 10:00:00    1.319       0.193     5.642          0.709     0.261
#> 84  2020-05-04 11:00:00    0.919       0.175     4.119          0.567     0.282
#> 85  2020-05-04 12:00:00    0.810       0.184     5.683          0.904     0.203
#> 86  2020-05-04 13:00:00    0.791       0.173     3.884          0.762     0.218
#> 87  2020-05-04 14:00:00    0.884       0.178     3.291          0.691     0.184
#> 88  2020-05-04 15:00:00    0.884       0.178     3.496          0.673     0.199
#> 89  2020-05-04 16:00:00    0.670       0.162     3.537          0.532     0.159
#> 90  2020-05-04 17:00:00    0.675       0.164     3.506          0.780     0.127
#> 91  2020-05-04 18:00:00    0.861       0.158     4.140          0.638     0.169
#> 92  2020-05-04 19:00:00    1.007       0.162     5.704          0.709     0.161
#> 93  2020-05-04 20:00:00    2.256       0.173     6.225          1.081     0.159
#> 94  2020-05-04 21:00:00    1.218       0.169     4.937          0.886     0.188
#> 95  2020-05-04 22:00:00    0.909       0.158     4.038          0.709     0.187
#> 96  2020-05-04 23:00:00    1.654       0.182     5.602          0.957     0.180
#> 97  2020-05-05 00:00:00    1.035       0.157     4.712          0.904     0.412
#> 98  2020-05-05 01:00:00    1.363       0.188     5.939          0.886     0.380
#> 99  2020-05-05 02:00:00    2.020       0.252     8.392          0.851     0.316
#> 100 2020-05-05 03:00:00    4.383       0.338    15.578          2.109     0.285
#> 101 2020-05-05 04:00:00    5.431       0.533    18.870          1.435     0.389
#> 102 2020-05-05 05:00:00    6.189       0.774    20.137          1.577     0.504
#> 103 2020-05-05 06:00:00    6.717       0.701    21.946          1.701     0.438
#> 104 2020-05-05 07:00:00    2.694       0.270     8.873          1.099     0.340
#> 105 2020-05-05 08:00:00    1.636       0.215     7.390          0.851     0.338
#> 106 2020-05-05 09:00:00    1.271       0.215     5.090          0.744     0.260
#> 107 2020-05-05 10:00:00    1.242       0.276     4.896          1.240     0.157
#> 108 2020-05-05 11:00:00    0.688       0.180     4.242          0.549     0.126
#> 109 2020-05-05 12:00:00    0.554       0.164     3.353          1.737     0.100
#> 110 2020-05-05 13:00:00    0.561       0.171     2.831          0.532     0.072
#> 111 2020-05-05 14:00:00    0.751       0.167     3.353          0.780     0.105
#> 112 2020-05-05 15:00:00    0.863       0.167     4.027          0.638     0.110
#> 113 2020-05-05 16:00:00    1.177       0.173     6.470          0.904     0.142
#> 114 2020-05-05 17:00:00    1.576       0.186     5.274          1.010     0.130
#> 115 2020-05-05 18:00:00    1.610       0.210     6.624          0.851     0.179
#> 116 2020-05-05 19:00:00    1.327       0.213     7.513          1.028     0.212
#> 117 2020-05-05 20:00:00    1.418       0.226    12.685          0.727     0.313
#> 118 2020-05-05 21:00:00    1.948       0.204    10.130          0.868     0.195
#> 119 2020-05-05 22:00:00    2.235       0.188    11.316          0.957     0.187
#> 120 2020-05-05 23:00:00    2.827       0.215    14.341          1.453     0.208
#> 
#> $OFP_Result_stat
#>          Species  Mean    SD   Min   Q25   Q50   Q75    Max Proportion
#> 3      Propylene 7.160 5.759 2.382 3.812 5.244 7.452 41.531     0.7119
#> 1       n.Butane 1.505 1.258 0.457 0.779 1.122 1.636  7.833     0.1496
#> 4 trans.2.Butene 0.957 0.484 0.355 0.673 0.798 0.997  3.562     0.0952
#> 2    Cyclohexane 0.223 0.089 0.145 0.171 0.196 0.246  0.774     0.0222
#> 5      Acetylene 0.213 0.095 0.058 0.150 0.193 0.258  0.504     0.0212
#> 
#> $OFP_Result_group
#>                    Time Alkanes Alkenes_exclude_BVOC Alkynes
#> 1   2020-05-01 00:00:00   0.981                2.736   0.090
#> 2   2020-05-01 01:00:00   1.209                4.340   0.104
#> 3   2020-05-01 02:00:00   1.438                5.943   0.119
#> 4   2020-05-01 03:00:00   1.894                9.150   0.147
#> 5   2020-05-01 04:00:00   2.085                8.055   0.150
#> 6   2020-05-01 05:00:00   1.502                6.368   0.191
#> 7   2020-05-01 06:00:00   1.447                6.175   0.205
#> 8   2020-05-01 07:00:00   1.404                6.789   0.242
#> 9   2020-05-01 08:00:00   1.305                6.971   0.284
#> 10  2020-05-01 09:00:00   1.532                5.491   0.267
#> 11  2020-05-01 10:00:00   1.132                4.011   0.115
#> 12  2020-05-01 11:00:00   0.919                4.933   0.058
#> 13  2020-05-01 12:00:00   0.860                4.189   0.077
#> 14  2020-05-01 13:00:00   0.998                4.310   0.087
#> 15  2020-05-01 14:00:00   0.840                4.016   0.088
#> 16  2020-05-01 15:00:00   0.866                5.050   0.149
#> 17  2020-05-01 16:00:00   1.124                5.500   0.146
#> 18  2020-05-01 17:00:00   1.057                4.539   0.139
#> 19  2020-05-01 18:00:00   0.898                6.630   0.139
#> 20  2020-05-01 19:00:00   1.619                8.893   0.205
#> 21  2020-05-01 20:00:00   1.929               12.100   0.311
#> 22  2020-05-01 21:00:00   1.849               11.740   0.294
#> 23  2020-05-01 22:00:00   1.977                8.848   0.188
#> 24  2020-05-01 23:00:00   1.550                7.389   0.296
#> 25  2020-05-02 00:00:00   1.040                6.114   0.176
#> 26  2020-05-02 01:00:00   0.879                7.647   0.130
#> 27  2020-05-02 02:00:00   0.995                6.633   0.109
#> 28  2020-05-02 03:00:00   1.228                4.605   0.066
#> 29  2020-05-02 04:00:00   1.362                5.232   0.105
#> 30  2020-05-02 05:00:00   1.629                6.485   0.183
#> 31  2020-05-02 06:00:00   1.539                7.943   0.256
#> 32  2020-05-02 07:00:00   1.837                7.041   0.499
#> 33  2020-05-02 08:00:00   1.782                7.474   0.454
#> 34  2020-05-02 09:00:00   1.673                8.340   0.362
#> 35  2020-05-02 10:00:00   1.687                8.140   0.331
#> 36  2020-05-02 11:00:00   2.200                9.158   0.303
#> 37  2020-05-02 12:00:00   1.895                7.838   0.338
#> 38  2020-05-02 13:00:00   1.029                4.910   0.183
#> 39  2020-05-02 14:00:00   0.708                4.084   0.096
#> 40  2020-05-02 15:00:00   0.617                3.711   0.088
#> 41  2020-05-02 16:00:00   0.646                3.508   0.097
#> 42  2020-05-02 17:00:00   1.005                4.265   0.113
#> 43  2020-05-02 18:00:00   1.167                5.091   0.166
#> 44  2020-05-02 19:00:00   1.663                7.442   0.154
#> 45  2020-05-02 20:00:00   2.712               16.620   0.348
#> 46  2020-05-02 21:00:00   2.540               17.775   0.270
#> 47  2020-05-02 22:00:00   2.434               12.531   0.216
#> 48  2020-05-02 23:00:00   4.252               18.439   0.179
#> 49  2020-05-03 00:00:00   4.183               18.896   0.183
#> 50  2020-05-03 01:00:00   4.044               19.812   0.191
#> 51  2020-05-03 02:00:00   3.767               21.643   0.207
#> 52  2020-05-03 03:00:00   8.100               45.093   0.249
#> 53  2020-05-03 04:00:00   4.283               22.409   0.225
#> 54  2020-05-03 05:00:00   3.812               24.311   0.290
#> 55  2020-05-03 06:00:00   3.308               25.572   0.490
#> 56  2020-05-03 07:00:00   2.312               13.268   0.246
#> 57  2020-05-03 08:00:00   1.354                5.140   0.160
#> 58  2020-05-03 09:00:00   1.457                5.504   0.191
#> 59  2020-05-03 10:00:00   0.922                4.796   0.170
#> 60  2020-05-03 11:00:00   0.949                4.045   0.156
#> 61  2020-05-03 12:00:00   0.859                4.699   0.149
#> 62  2020-05-03 13:00:00   0.790                3.456   0.165
#> 63  2020-05-03 14:00:00   0.710                3.736   0.175
#> 64  2020-05-03 15:00:00   0.799                3.636   0.159
#> 65  2020-05-03 16:00:00   1.218                4.394   0.186
#> 66  2020-05-03 17:00:00   0.929                4.854   0.228
#> 67  2020-05-03 18:00:00   1.350                5.506   0.236
#> 68  2020-05-03 19:00:00   1.015                6.169   0.257
#> 69  2020-05-03 20:00:00   1.129                6.520   0.244
#> 70  2020-05-03 21:00:00   0.939                6.011   0.223
#> 71  2020-05-03 22:00:00   1.015                4.288   0.206
#> 72  2020-05-03 23:00:00   0.858                4.159   0.202
#> 73  2020-05-04 00:00:00   0.875                4.368   0.229
#> 74  2020-05-04 01:00:00   0.887                4.173   0.223
#> 75  2020-05-04 02:00:00   0.910                3.783   0.211
#> 76  2020-05-04 03:00:00   0.957                4.118   0.233
#> 77  2020-05-04 04:00:00   0.792                3.639   0.215
#> 78  2020-05-04 05:00:00   1.043                4.391   0.357
#> 79  2020-05-04 06:00:00   1.186                5.124   0.329
#> 80  2020-05-04 07:00:00   1.391                5.565   0.255
#> 81  2020-05-04 08:00:00   1.459                5.559   0.251
#> 82  2020-05-04 09:00:00   1.629                6.400   0.250
#> 83  2020-05-04 10:00:00   1.512                6.351   0.261
#> 84  2020-05-04 11:00:00   1.094                4.686   0.282
#> 85  2020-05-04 12:00:00   0.995                6.587   0.203
#> 86  2020-05-04 13:00:00   0.964                4.646   0.218
#> 87  2020-05-04 14:00:00   1.062                3.983   0.184
#> 88  2020-05-04 15:00:00   1.062                4.169   0.199
#> 89  2020-05-04 16:00:00   0.832                4.068   0.159
#> 90  2020-05-04 17:00:00   0.840                4.286   0.127
#> 91  2020-05-04 18:00:00   1.018                4.778   0.169
#> 92  2020-05-04 19:00:00   1.169                6.413   0.161
#> 93  2020-05-04 20:00:00   2.429                7.306   0.159
#> 94  2020-05-04 21:00:00   1.387                5.823   0.188
#> 95  2020-05-04 22:00:00   1.067                4.746   0.187
#> 96  2020-05-04 23:00:00   1.836                6.559   0.180
#> 97  2020-05-05 00:00:00   1.191                5.616   0.412
#> 98  2020-05-05 01:00:00   1.552                6.825   0.380
#> 99  2020-05-05 02:00:00   2.272                9.243   0.316
#> 100 2020-05-05 03:00:00   4.721               17.687   0.285
#> 101 2020-05-05 04:00:00   5.963               20.305   0.389
#> 102 2020-05-05 05:00:00   6.963               21.714   0.504
#> 103 2020-05-05 06:00:00   7.419               23.648   0.438
#> 104 2020-05-05 07:00:00   2.964                9.971   0.340
#> 105 2020-05-05 08:00:00   1.851                8.241   0.338
#> 106 2020-05-05 09:00:00   1.486                5.835   0.260
#> 107 2020-05-05 10:00:00   1.518                6.137   0.157
#> 108 2020-05-05 11:00:00   0.868                4.791   0.126
#> 109 2020-05-05 12:00:00   0.719                5.089   0.100
#> 110 2020-05-05 13:00:00   0.732                3.363   0.072
#> 111 2020-05-05 14:00:00   0.917                4.133   0.105
#> 112 2020-05-05 15:00:00   1.030                4.665   0.110
#> 113 2020-05-05 16:00:00   1.350                7.374   0.142
#> 114 2020-05-05 17:00:00   1.763                6.285   0.130
#> 115 2020-05-05 18:00:00   1.820                7.474   0.179
#> 116 2020-05-05 19:00:00   1.540                8.541   0.212
#> 117 2020-05-05 20:00:00   1.643               13.412   0.313
#> 118 2020-05-05 21:00:00   2.152               10.998   0.195
#> 119 2020-05-05 22:00:00   2.423               12.273   0.187
#> 120 2020-05-05 23:00:00   3.041               15.794   0.208
#> 
#> $OFP_Result_group_stat
#>                Species  Mean    SD   Min   Q25   Q50   Q75    Max Proportion
#> 2 Alkenes_exclude_BVOC 8.117 6.172 2.736 4.588 6.125 8.266 45.093     0.8070
#> 1              Alkanes 1.728 1.320 0.617 0.977 1.352 1.840  8.100     0.1718
#> 3              Alkynes 0.213 0.095 0.058 0.150 0.193 0.258  0.504     0.0212
#>