Calculate TUV in batch

This function runs TUV in batch by reading the time series for the parameters to be entered, and summarizes the results to the new dataframe.

Usage

tuv(pathtuv, df, colid = 1)

Arguments

pathtuv

path for parent folder of TUV executable for Windows, such as "c:/tuv5.3.1.exe".

df

dataframe of the time series for the parameters to be entered, such as 'date', 'o3col'. It must includes date column.

colid

column index of date. The default value is 1.

Value

a dataframe.The first column is datetime. The second column is the solar altitude Angle. The rates of photolysis for each reaction(Unit: s-1) start from third column: 1 = O3 -> O2 + O1D

Details

There are online and offline versions of the TUV model, but both need to run on a daily basis (that means manually reset parameters for each day's simulation).
This function runs TUV in batch by reading the time series for the parameters to be entered, and summarizes the results to the new dataframe.
Currently only mode 2 (mode that outputs the photolysis rates) is supported.
Logical variables are not supported currently!!!
Please download TUV executable for Windows before you use this function.

Columns of photolysis rate coefficients (s-1):
1 = O3 -> O2 + O(1D)
2 = H2O2 -> 2 OH
3 = NO2 -> NO + O(3P)
4 = NO3 -> NO + O2
5 = NO3 -> NO2 + O(3P)
6 = CH2O -> H + HCO
7 = CH2O -> H2 + CO