Convertion and analysis of VOC concentrations

convert unit of VOCs between micrograms per cubic meter (ugm) and parts per billion by volume (ppbv); conduct statistics of VOC concentrations. Note: for Chinese VOC name, please also use English punctuation.

Usage

vocct(
  df,
  unit = "ppbv",
  t = 25,
  p = 101.325,
  stcd = FALSE,
  sortd = TRUE,
  chn = FALSE,
  bvoc = TRUE
)

Arguments

df

dataframe contains time series.

unit

unit for VOC concentration. A character vector from these options: "ugm" or "ppbv". "ugm" means ug/m3. "ppbv" means part per billion volumn.

t

Temperature, in Degrees Celsius, used to convert data in micrograms per cubic meter to standard conditions (25 Degrees Celsius, 101.325 kPa). By default, t equals to 25 Degrees Celsius.

p

Pressure, in kPa, used to convert data in micrograms per cubic meter to standard conditions (25 Degrees Celsius, 101.325 kPa). By default, p equals to 101.325 kPa.

stcd

logical. Does it output results in standard conditions? The default vaule is FALSE.

sortd

logical value. It determines whether the VOC species are sorted or not. By default, sortd has value "TRUE". If TRUE, VOC species in time series will be arranged according to VOC group, relative molecular weight, and MIR value.

chn

logical. Dose colnames present as Chinese? The default vaule is FALSE.

bvoc

logical. Whether you want to list BVOC as a separate VOC group? The default vaule is TRUE.

Value

a list contains 9 tables: MW_Result: the matched Molecular Weight (MW) value result; Con_ugm: the time series of VOC mass concentration by species; Con_ugm_stat: the statistics of VOC mass concentration by species; Con_ugm_group: the time series of VOC mass concentration classified by groups; Con_ugm_group_stat: the statistics of VOC mass concentrationaccording to major groups; Con_ppbv: time series of VOC volume concentration by species; Con_ppbv_stat: the statistics of VOC volume concentration by species; Con_ppbv_group: the time series of VOC volume concentration according to major groups; Con_ppbv_group_stat: the time series of VOC volume concentration classified by groups.

Details

The CAS number was matched for each VOC speices (from column name), and the Molecular Weight (MW) value and Maximum Incremental Reactivity (MIR) value are matched through the CAS number and used for time series calculation.
The MIR value comes from "Carter, W. P. (2009). Updated maximum incremental reactivity scale and hydrocarbon bin reactivities for regulatory applications. California Air Resources Board Contract, 2009, 339" (revised January 28, 2010).

Examples

vocct(voc)
#> $MW_Result
#>             Name      CAS   Matched_Name   MIR    MW   Group Raw_order
#> 3       n-Butane 106-97-8       n-butane  1.15 58.12 Alkanes         3
#> 5    Cyclohexane 110-82-7    cyclohexane  1.25 84.16 Alkanes         5
#> 1      Propylene 115-07-1        propene 11.66 42.08 Alkenes         1
#> 4 trans-2-Butene 624-64-6 trans-2-butene 15.16 56.11 Alkenes         4
#> 2      Acetylene  74-86-2      acetylene  0.95 26.04 Alkynes         2
#> 
#> $Con_ugm
#>                    Time   n.Butane Cyclohexane Propylene trans.2.Butene
#> 1   2020-05-01 00:00:00  1.2931321   0.3510947 0.4010052     0.04589746
#> 2   2020-05-01 01:00:00  1.6734650   0.3597000 0.6471157     0.06425644
#> 3   2020-05-01 02:00:00  2.0537980   0.3683052 0.8932262     0.08261543
#> 4   2020-05-01 03:00:00  2.8144639   0.3855157 1.3854472     0.11933339
#> 5   2020-05-01 04:00:00  3.0997136   0.4233789 1.1324525     0.17211547
#> 6   2020-05-01 05:00:00  2.1488812   0.3820736 0.9259262     0.11244877
#> 7   2020-05-01 06:00:00  1.9919939   0.4405894 0.9173210     0.09408979
#> 8   2020-05-01 07:00:00  1.8731398   0.4818947 0.9878841     0.11933339
#> 9   2020-05-01 08:00:00  1.7828107   0.4096105 1.0722157     0.07802568
#> 10  2020-05-01 09:00:00  2.2487186   0.3373263 0.7693104     0.11933339
#> 11  2020-05-01 10:00:00  1.6259234   0.2822526 0.5679473     0.08261543
#> 12  2020-05-01 11:00:00  1.2693612   0.2753684 0.7021894     0.09867954
#> 13  2020-05-01 12:00:00  1.1766551   0.2684842 0.5799947     0.09638466
#> 14  2020-05-01 13:00:00  1.3787070   0.2994631 0.5765526     0.11474365
#> 15  2020-05-01 14:00:00  1.1196051   0.2891368 0.5627841     0.08720517
#> 16  2020-05-01 15:00:00  1.1647697   0.2891368 0.7280052     0.09408979
#> 17  2020-05-01 16:00:00  1.5712505   0.3201158 0.8037315     0.09408979
#> 18  2020-05-01 17:00:00  1.4785444   0.2994631 0.6746526     0.06884619
#> 19  2020-05-01 18:00:00  1.2194425   0.2891368 0.9999315     0.08950004
#> 20  2020-05-01 19:00:00  2.4198684   0.3166736 1.3630735     0.10326928
#> 21  2020-05-01 20:00:00  2.9261867   0.3373263 1.9000419     0.10556415
#> 22  2020-05-01 21:00:00  2.8002014   0.3270000 1.7589156     0.16752572
#> 23  2020-05-01 22:00:00  3.0117616   0.3338842 1.3166051     0.13310263
#> 24  2020-05-01 23:00:00  2.2606040   0.3545368 1.1186841     0.09638466
#> 25  2020-05-02 00:00:00  1.3858382   0.3579789 0.9190420     0.08491030
#> 26  2020-05-02 01:00:00  0.9853001   0.4732894 1.1845143     0.07917312
#> 27  2020-05-02 02:00:00  1.1861634   0.4715684 1.0033736     0.08720517
#> 28  2020-05-02 03:00:00  1.5878901   0.4681263 0.6410920     0.10326928
#> 29  2020-05-02 04:00:00  1.7899420   0.4922210 0.7555420     0.09638466
#> 30  2020-05-02 05:00:00  2.1940458   0.5404105 0.9844420     0.08261543
#> 31  2020-05-02 06:00:00  2.1179792   0.4681263 1.1582683     0.13769238
#> 32  2020-05-02 07:00:00  2.6944213   0.4061684 0.9422762     0.18703214
#> 33  2020-05-02 08:00:00  2.5601162   0.4440315 1.0584473     0.15375649
#> 34  2020-05-02 09:00:00  2.2915061   0.5197578 1.2907893     0.08720517
#> 35  2020-05-02 10:00:00  2.3105227   0.5231999 1.2512051     0.09179492
#> 36  2020-05-02 11:00:00  3.1306157   0.5748315 1.4284735     0.08720517
#> 37  2020-05-02 12:00:00  2.7930702   0.4061684 1.1823630     0.10556415
#> 38  2020-05-02 13:00:00  1.4191173   0.3097894 0.7073526     0.09179492
#> 39  2020-05-02 14:00:00  0.9318157   0.2547158 0.5593420     0.09867954
#> 40  2020-05-02 15:00:00  0.7796826   0.2512737 0.4904999     0.10326928
#> 41  2020-05-02 16:00:00  0.8105846   0.2684842 0.4922210     0.07573081
#> 42  2020-05-02 17:00:00  1.4357569   0.2581579 0.5868789     0.10097441
#> 43  2020-05-02 18:00:00  1.6924817   0.2753684 0.7228420     0.10326928
#> 44  2020-05-02 19:00:00  2.5101975   0.3029052 1.0859841     0.12851288
#> 45  2020-05-02 20:00:00  4.2668604   0.3338842 2.5953471     0.15605136
#> 46  2020-05-02 21:00:00  3.9839877   0.3235579 2.7003313     0.22489755
#> 47  2020-05-02 22:00:00  3.7652963   0.3579789 1.9516735     0.12162826
#> 48  2020-05-02 23:00:00  6.8269766   0.3975631 2.6740852     0.33103542
#> 49  2020-05-03 00:00:00  6.7104996   0.3958421 2.7683129     0.31783990
#> 50  2020-05-03 01:00:00  6.4775457   0.3924000 2.9567681     0.29144886
#> 51  2020-05-03 02:00:00  6.0116378   0.3855157 3.3336786     0.23866678
#> 52  2020-05-03 03:00:00 13.3710806   0.4199368 6.9926361     0.46126946
#> 53  2020-05-03 04:00:00  6.8959119   0.3820736 3.4627575     0.23866678
#> 54  2020-05-03 05:00:00  6.0544253   0.4164947 3.8396680     0.19506420
#> 55  2020-05-03 06:00:00  5.1202324   0.4853368 3.8878575     0.32128221
#> 56  2020-05-03 07:00:00  3.5323423   0.3820736 2.0101893     0.17211547
#> 57  2020-05-03 08:00:00  1.9777314   0.3063473 0.7520999     0.08720517
#> 58  2020-05-03 09:00:00  2.1726520   0.2891368 0.7865210     0.10785903
#> 59  2020-05-03 10:00:00  1.2741154   0.2753684 0.6970262     0.08491030
#> 60  2020-05-03 11:00:00  1.3287883   0.2684842 0.5318052     0.11474365
#> 61  2020-05-03 12:00:00  1.1814093   0.2616000 0.6419526     0.11474365
#> 62  2020-05-03 13:00:00  1.0673094   0.2581579 0.4715684     0.08491030
#> 63  2020-05-03 14:00:00  0.9318157   0.2581579 0.4887789     0.10785903
#> 64  2020-05-03 15:00:00  1.1172281   0.2271789 0.5077105     0.08032055
#> 65  2020-05-03 16:00:00  1.7946961   0.2616000 0.6144157     0.09638466
#> 66  2020-05-03 17:00:00  1.2907550   0.2719263 0.7039104     0.08720517
#> 67  2020-05-03 18:00:00  2.0205188   0.2616000 0.8106157     0.08950004
#> 68  2020-05-03 19:00:00  1.4452652   0.2650421 0.8983894     0.10785903
#> 69  2020-05-03 20:00:00  1.6092838   0.2925789 0.9844420     0.08720517
#> 70  2020-05-03 21:00:00  1.3145258   0.2650421 0.8450367     0.12851288
#> 71  2020-05-03 22:00:00  1.4476423   0.2616000 0.6264631     0.07343593
#> 72  2020-05-03 23:00:00  1.1909176   0.2512737 0.5868789     0.08720517
#> 73  2020-05-04 00:00:00  1.2194425   0.2529947 0.5817157     0.11818596
#> 74  2020-05-04 01:00:00  1.2337050   0.2581579 0.5593420     0.11015390
#> 75  2020-05-04 02:00:00  1.2622300   0.2684842 0.5145947     0.09408979
#> 76  2020-05-04 03:00:00  1.3311654   0.2788105 0.5799947     0.08720517
#> 77  2020-05-04 04:00:00  1.0673094   0.2616000 0.5231999     0.06884619
#> 78  2020-05-04 05:00:00  1.4666590   0.2891368 0.6109736     0.09867954
#> 79  2020-05-04 06:00:00  1.6687109   0.3270000 0.6866999     0.13539750
#> 80  2020-05-04 07:00:00  1.9587147   0.3820736 0.7847999     0.11703852
#> 81  2020-05-04 08:00:00  2.0942084   0.3648631 0.8226631     0.08720517
#> 82  2020-05-04 09:00:00  2.4293768   0.3235579 0.8777367     0.15375649
#> 83  2020-05-04 10:00:00  2.2510957   0.3029052 0.9500209     0.09179492
#> 84  2020-05-04 11:00:00  1.5688735   0.2753684 0.6935841     0.07343593
#> 85  2020-05-04 12:00:00  1.3834611   0.2891368 0.9569052     0.11703852
#> 86  2020-05-04 13:00:00  1.3501820   0.2719263 0.6539999     0.09867954
#> 87  2020-05-04 14:00:00  1.5094464   0.2788105 0.5541789     0.08950004
#> 88  2020-05-04 15:00:00  1.5094464   0.2788105 0.5885999     0.08720517
#> 89  2020-05-04 16:00:00  1.1433760   0.2547158 0.5954841     0.06884619
#> 90  2020-05-04 17:00:00  1.1528843   0.2581579 0.5903210     0.10097441
#> 91  2020-05-04 18:00:00  1.4690361   0.2478316 0.6970262     0.08261543
#> 92  2020-05-04 19:00:00  1.7186296   0.2547158 0.9603473     0.09179492
#> 93  2020-05-04 20:00:00  3.8508712   0.2719263 1.0481209     0.13998725
#> 94  2020-05-04 21:00:00  2.0799459   0.2650421 0.8312683     0.11474365
#> 95  2020-05-04 22:00:00  1.5522339   0.2478316 0.6798157     0.09179492
#> 96  2020-05-04 23:00:00  2.8239722   0.2856947 0.9431367     0.12392314
#> 97  2020-05-05 00:00:00  1.7661712   0.2461105 0.7934052     0.11703852
#> 98  2020-05-05 01:00:00  2.3271623   0.2960210 0.9999315     0.11474365
#> 99  2020-05-05 02:00:00  3.4491445   0.3958421 1.4129841     0.11015390
#> 100 2020-05-05 03:00:00  7.4830509   0.5300842 2.6228839     0.27308988
#> 101 2020-05-05 04:00:00  9.2706159   0.8364315 3.1770628     0.18588471
#> 102 2020-05-05 05:00:00 10.5661250   1.2150630 3.3904733     0.20424369
#> 103 2020-05-05 06:00:00 11.4670387   1.1014736 3.6950996     0.22030780
#> 104 2020-05-05 07:00:00  4.5996517   0.4233789 1.4938735     0.14228212
#> 105 2020-05-05 08:00:00  2.7930702   0.3373263 1.2443209     0.11015390
#> 106 2020-05-05 09:00:00  2.1702749   0.3373263 0.8570841     0.09638466
#> 107 2020-05-05 10:00:00  2.1203562   0.4337052 0.8243841     0.16064110
#> 108 2020-05-05 11:00:00  1.1742780   0.2822526 0.7142368     0.07114106
#> 109 2020-05-05 12:00:00  0.9460782   0.2581579 0.5645052     0.22489755
#> 110 2020-05-05 13:00:00  0.9579636   0.2684842 0.4767315     0.06884619
#> 111 2020-05-05 14:00:00  1.2812467   0.2616000 0.5645052     0.10097441
#> 112 2020-05-05 15:00:00  1.4737902   0.2616000 0.6780947     0.08261543
#> 113 2020-05-05 16:00:00  2.0086334   0.2719263 1.0894262     0.11703852
#> 114 2020-05-05 17:00:00  2.6908557   0.2925789 0.8880631     0.13080776
#> 115 2020-05-05 18:00:00  2.7479056   0.3304421 1.1152420     0.11015390
#> 116 2020-05-05 19:00:00  2.2653582   0.3338842 1.2649735     0.13310263
#> 117 2020-05-05 20:00:00  2.4198684   0.3545368 2.1358261     0.09408979
#> 118 2020-05-05 21:00:00  3.3255363   0.3201158 1.7055630     0.11244877
#> 119 2020-05-05 22:00:00  3.8152150   0.2960210 1.9052051     0.12392314
#> 120 2020-05-05 23:00:00  4.8254744   0.3373263 2.4146366     0.18817958
#>     Acetylene
#> 1   0.1863792
#> 2   0.2156673
#> 3   0.2449555
#> 4   0.3035318
#> 5   0.3109870
#> 6   0.3940588
#> 7   0.4228145
#> 8   0.5005612
#> 9   0.5868282
#> 10  0.5527474
#> 11  0.2375003
#> 12  0.1192827
#> 13  0.1597536
#> 14  0.1799890
#> 15  0.1810541
#> 16  0.3088569
#> 17  0.3024668
#> 18  0.2875565
#> 19  0.2875565
#> 20  0.4238795
#> 21  0.6422094
#> 22  0.6070636
#> 23  0.3876687
#> 24  0.6113237
#> 25  0.3642382
#> 26  0.2689185
#> 27  0.2247200
#> 28  0.1363231
#> 29  0.2172649
#> 30  0.3791485
#> 31  0.5293169
#> 32  1.0314756
#> 33  0.9372210
#> 34  0.7487118
#> 35  0.6848104
#> 36  0.6262341
#> 37  0.6986557
#> 38  0.3791485
#> 39  0.1991595
#> 40  0.1821191
#> 41  0.2002245
#> 42  0.2343053
#> 43  0.3429377
#> 44  0.3184421
#> 45  0.7188911
#> 46  0.5580725
#> 47  0.4473100
#> 48  0.3703621
#> 49  0.3786160
#> 50  0.3951239
#> 51  0.4281396
#> 52  0.5154716
#> 53  0.4654154
#> 54  0.5985434
#> 55  1.0117727
#> 56  0.5090814
#> 57  0.3301574
#> 58  0.3940588
#> 59  0.3514579
#> 60  0.3216372
#> 61  0.3088569
#> 62  0.3408076
#> 63  0.3621081
#> 64  0.3290924
#> 65  0.3834086
#> 66  0.4718056
#> 67  0.4877809
#> 68  0.5314469
#> 69  0.5037563
#> 70  0.4600903
#> 71  0.4249445
#> 72  0.4164243
#> 73  0.4739356
#> 74  0.4611553
#> 75  0.4355948
#> 76  0.4824558
#> 77  0.4451800
#> 78  0.7369965
#> 79  0.6794852
#> 80  0.5261218
#> 81  0.5186666
#> 82  0.5165366
#> 83  0.5389021
#> 84  0.5836331
#> 85  0.4185544
#> 86  0.4505051
#> 87  0.3802135
#> 88  0.4110992
#> 89  0.3290924
#> 90  0.2630609
#> 91  0.3493278
#> 92  0.3333525
#> 93  0.3290924
#> 94  0.3876687
#> 95  0.3866037
#> 96  0.3716933
#> 97  0.8504216
#> 98  0.7849226
#> 99  0.6539247
#> 100 0.5889582
#> 101 0.8030280
#> 102 1.0405283
#> 103 0.9052703
#> 104 0.7018507
#> 105 0.6975907
#> 106 0.5378371
#> 107 0.3248323
#> 108 0.2609309
#> 109 0.2076797
#> 110 0.1491033
#> 111 0.2161999
#> 112 0.2279151
#> 113 0.2928816
#> 114 0.2694510
#> 115 0.3706283
#> 116 0.4377248
#> 117 0.6464695
#> 118 0.4036441
#> 119 0.3855387
#> 120 0.4302697
#> 
#> $Con_ugm_stat
#>          Species  Mean    SD   Min   Q25   Q50   Q75    Max Proportion
#> 1       n.Butane 2.569 2.147 0.780 1.331 1.916 2.793 13.371     0.5478
#> 3      Propylene 1.206 0.970 0.401 0.642 0.883 1.255  6.993     0.2571
#> 5      Acetylene 0.441 0.196 0.119 0.310 0.399 0.533  1.041     0.0940
#> 2    Cyclohexane 0.350 0.139 0.227 0.268 0.308 0.386  1.215     0.0746
#> 4 trans.2.Butene 0.124 0.063 0.046 0.087 0.103 0.129  0.461     0.0264
#> 
#> $Con_ugm_group
#>                    Time   Alkanes Alkenes_exclude_BVOC   Alkynes
#> 1   2020-05-01 00:00:00  1.644227            0.4469027 0.1863792
#> 2   2020-05-01 01:00:00  2.033165            0.7113722 0.2156673
#> 3   2020-05-01 02:00:00  2.422103            0.9758416 0.2449555
#> 4   2020-05-01 03:00:00  3.199980            1.5047806 0.3035318
#> 5   2020-05-01 04:00:00  3.523093            1.3045680 0.3109870
#> 6   2020-05-01 05:00:00  2.530955            1.0383750 0.3940588
#> 7   2020-05-01 06:00:00  2.432583            1.0114107 0.4228145
#> 8   2020-05-01 07:00:00  2.355035            1.1072175 0.5005612
#> 9   2020-05-01 08:00:00  2.192421            1.1502414 0.5868282
#> 10  2020-05-01 09:00:00  2.586045            0.8886438 0.5527474
#> 11  2020-05-01 10:00:00  1.908176            0.6505627 0.2375003
#> 12  2020-05-01 11:00:00  1.544730            0.8008689 0.1192827
#> 13  2020-05-01 12:00:00  1.445139            0.6763793 0.1597536
#> 14  2020-05-01 13:00:00  1.678170            0.6912962 0.1799890
#> 15  2020-05-01 14:00:00  1.408742            0.6499893 0.1810541
#> 16  2020-05-01 15:00:00  1.453906            0.8220950 0.3088569
#> 17  2020-05-01 16:00:00  1.891366            0.8978213 0.3024668
#> 18  2020-05-01 17:00:00  1.778007            0.7434987 0.2875565
#> 19  2020-05-01 18:00:00  1.508579            1.0894315 0.2875565
#> 20  2020-05-01 19:00:00  2.736542            1.4663428 0.4238795
#> 21  2020-05-01 20:00:00  3.263513            2.0056061 0.6422094
#> 22  2020-05-01 21:00:00  3.127201            1.9264413 0.6070636
#> 23  2020-05-01 22:00:00  3.345646            1.4497077 0.3876687
#> 24  2020-05-01 23:00:00  2.615141            1.2150688 0.6113237
#> 25  2020-05-02 00:00:00  1.743817            1.0039523 0.3642382
#> 26  2020-05-02 01:00:00  1.458589            1.2636875 0.2689185
#> 27  2020-05-02 02:00:00  1.657732            1.0905787 0.2247200
#> 28  2020-05-02 03:00:00  2.056016            0.7443613 0.1363231
#> 29  2020-05-02 04:00:00  2.282163            0.8519267 0.2172649
#> 30  2020-05-02 05:00:00  2.734456            1.0670574 0.3791485
#> 31  2020-05-02 06:00:00  2.586105            1.2959607 0.5293169
#> 32  2020-05-02 07:00:00  3.100590            1.1293084 1.0314756
#> 33  2020-05-02 08:00:00  3.004148            1.2122037 0.9372210
#> 34  2020-05-02 09:00:00  2.811264            1.3779945 0.7487118
#> 35  2020-05-02 10:00:00  2.833723            1.3430000 0.6848104
#> 36  2020-05-02 11:00:00  3.705447            1.5156787 0.6262341
#> 37  2020-05-02 12:00:00  3.199239            1.2879272 0.6986557
#> 38  2020-05-02 13:00:00  1.728907            0.7991475 0.3791485
#> 39  2020-05-02 14:00:00  1.186532            0.6580216 0.1991595
#> 40  2020-05-02 15:00:00  1.030956            0.5937692 0.1821191
#> 41  2020-05-02 16:00:00  1.079069            0.5679518 0.2002245
#> 42  2020-05-02 17:00:00  1.693915            0.6878533 0.2343053
#> 43  2020-05-02 18:00:00  1.967850            0.8261113 0.3429377
#> 44  2020-05-02 19:00:00  2.813103            1.2144970 0.3184421
#> 45  2020-05-02 20:00:00  4.600745            2.7513984 0.7188911
#> 46  2020-05-02 21:00:00  4.307546            2.9252288 0.5580725
#> 47  2020-05-02 22:00:00  4.123275            2.0733017 0.4473100
#> 48  2020-05-02 23:00:00  7.224540            3.0051207 0.3703621
#> 49  2020-05-03 00:00:00  7.106342            3.0861528 0.3786160
#> 50  2020-05-03 01:00:00  6.869946            3.2482170 0.3951239
#> 51  2020-05-03 02:00:00  6.397154            3.5723454 0.4281396
#> 52  2020-05-03 03:00:00 13.791017            7.4539055 0.5154716
#> 53  2020-05-03 04:00:00  7.277986            3.7014243 0.4654154
#> 54  2020-05-03 05:00:00  6.470920            4.0347322 0.5985434
#> 55  2020-05-03 06:00:00  5.605569            4.2091397 1.0117727
#> 56  2020-05-03 07:00:00  3.914416            2.1823047 0.5090814
#> 57  2020-05-03 08:00:00  2.284079            0.8393051 0.3301574
#> 58  2020-05-03 09:00:00  2.461789            0.8943800 0.3940588
#> 59  2020-05-03 10:00:00  1.549484            0.7819365 0.3514579
#> 60  2020-05-03 11:00:00  1.597272            0.6465489 0.3216372
#> 61  2020-05-03 12:00:00  1.443009            0.7566962 0.3088569
#> 62  2020-05-03 13:00:00  1.325467            0.5564787 0.3408076
#> 63  2020-05-03 14:00:00  1.189974            0.5966379 0.3621081
#> 64  2020-05-03 15:00:00  1.344407            0.5880310 0.3290924
#> 65  2020-05-03 16:00:00  2.056296            0.7108004 0.3834086
#> 66  2020-05-03 17:00:00  1.562681            0.7911156 0.4718056
#> 67  2020-05-03 18:00:00  2.282119            0.9001157 0.4877809
#> 68  2020-05-03 19:00:00  1.710307            1.0062484 0.5314469
#> 69  2020-05-03 20:00:00  1.901863            1.0716472 0.5037563
#> 70  2020-05-03 21:00:00  1.579568            0.9735496 0.4600903
#> 71  2020-05-03 22:00:00  1.709242            0.6998990 0.4249445
#> 72  2020-05-03 23:00:00  1.442191            0.6740841 0.4164243
#> 73  2020-05-04 00:00:00  1.472437            0.6999017 0.4739356
#> 74  2020-05-04 01:00:00  1.491863            0.6694959 0.4611553
#> 75  2020-05-04 02:00:00  1.530714            0.6086845 0.4355948
#> 76  2020-05-04 03:00:00  1.609976            0.6671998 0.4824558
#> 77  2020-05-04 04:00:00  1.328909            0.5920461 0.4451800
#> 78  2020-05-04 05:00:00  1.755796            0.7096532 0.7369965
#> 79  2020-05-04 06:00:00  1.995711            0.8220974 0.6794852
#> 80  2020-05-04 07:00:00  2.340788            0.9018384 0.5261218
#> 81  2020-05-04 08:00:00  2.459071            0.9098682 0.5186666
#> 82  2020-05-04 09:00:00  2.752935            1.0314932 0.5165366
#> 83  2020-05-04 10:00:00  2.554001            1.0418159 0.5389021
#> 84  2020-05-04 11:00:00  1.844242            0.7670201 0.5836331
#> 85  2020-05-04 12:00:00  1.672598            1.0739437 0.4185544
#> 86  2020-05-04 13:00:00  1.622108            0.7526795 0.4505051
#> 87  2020-05-04 14:00:00  1.788257            0.6436789 0.3802135
#> 88  2020-05-04 15:00:00  1.788257            0.6758051 0.4110992
#> 89  2020-05-04 16:00:00  1.398092            0.6643303 0.3290924
#> 90  2020-05-04 17:00:00  1.411042            0.6912954 0.2630609
#> 91  2020-05-04 18:00:00  1.716868            0.7796417 0.3493278
#> 92  2020-05-04 19:00:00  1.973345            1.0521422 0.3333525
#> 93  2020-05-04 20:00:00  4.122797            1.1881082 0.3290924
#> 94  2020-05-04 21:00:00  2.344988            0.9460120 0.3876687
#> 95  2020-05-04 22:00:00  1.800065            0.7716106 0.3866037
#> 96  2020-05-04 23:00:00  3.109667            1.0670599 0.3716933
#> 97  2020-05-05 00:00:00  2.012282            0.9104437 0.8504216
#> 98  2020-05-05 01:00:00  2.623183            1.1146751 0.7849226
#> 99  2020-05-05 02:00:00  3.844987            1.5231380 0.6539247
#> 100 2020-05-05 03:00:00  8.013135            2.8959738 0.5889582
#> 101 2020-05-05 04:00:00 10.107047            3.3629475 0.8030280
#> 102 2020-05-05 05:00:00 11.781188            3.5947170 1.0405283
#> 103 2020-05-05 06:00:00 12.568512            3.9154074 0.9052703
#> 104 2020-05-05 07:00:00  5.023031            1.6361556 0.7018507
#> 105 2020-05-05 08:00:00  3.130396            1.3544748 0.6975907
#> 106 2020-05-05 09:00:00  2.507601            0.9534688 0.5378371
#> 107 2020-05-05 10:00:00  2.554061            0.9850252 0.3248323
#> 108 2020-05-05 11:00:00  1.456531            0.7853778 0.2609309
#> 109 2020-05-05 12:00:00  1.204236            0.7894027 0.2076797
#> 110 2020-05-05 13:00:00  1.226448            0.5455777 0.1491033
#> 111 2020-05-05 14:00:00  1.542847            0.6654796 0.2161999
#> 112 2020-05-05 15:00:00  1.735390            0.7607101 0.2279151
#> 113 2020-05-05 16:00:00  2.280560            1.2064647 0.2928816
#> 114 2020-05-05 17:00:00  2.983435            1.0188708 0.2694510
#> 115 2020-05-05 18:00:00  3.078348            1.2253959 0.3706283
#> 116 2020-05-05 19:00:00  2.599242            1.3980762 0.4377248
#> 117 2020-05-05 20:00:00  2.774405            2.2299159 0.6464695
#> 118 2020-05-05 21:00:00  3.645652            1.8180117 0.4036441
#> 119 2020-05-05 22:00:00  4.111236            2.0291282 0.3855387
#> 120 2020-05-05 23:00:00  5.162801            2.6028162 0.4302697
#> 
#> $Con_ugm_group_stat
#>                Species  Mean    SD   min   Q25   Q50   Q75    Max Proportion
#> 1              Alkanes 2.919 2.244 1.031 1.639 2.282 3.114 13.791     0.6225
#> 2 Alkenes_exclude_BVOC 1.329 1.023 0.447 0.744 0.994 1.360  7.454     0.2834
#> 3              Alkynes 0.441 0.196 0.119 0.310 0.399 0.533  1.041     0.0940
#> 
#> $Con_ppbv
#>                    Time n.Butane Cyclohexane Propylene trans.2.Butene Acetylene
#> 1   2020-05-01 00:00:00   0.5440      0.1020   0.23300        0.02000   0.17500
#> 2   2020-05-01 01:00:00   0.7040      0.1045   0.37600        0.02800   0.20250
#> 3   2020-05-01 02:00:00   0.8640      0.1070   0.51900        0.03600   0.23000
#> 4   2020-05-01 03:00:00   1.1840      0.1120   0.80500        0.05200   0.28500
#> 5   2020-05-01 04:00:00   1.3040      0.1230   0.65800        0.07500   0.29200
#> 6   2020-05-01 05:00:00   0.9040      0.1110   0.53800        0.04900   0.37000
#> 7   2020-05-01 06:00:00   0.8380      0.1280   0.53300        0.04100   0.39700
#> 8   2020-05-01 07:00:00   0.7880      0.1400   0.57400        0.05200   0.47000
#> 9   2020-05-01 08:00:00   0.7500      0.1190   0.62300        0.03400   0.55100
#> 10  2020-05-01 09:00:00   0.9460      0.0980   0.44700        0.05200   0.51900
#> 11  2020-05-01 10:00:00   0.6840      0.0820   0.33000        0.03600   0.22300
#> 12  2020-05-01 11:00:00   0.5340      0.0800   0.40800        0.04300   0.11200
#> 13  2020-05-01 12:00:00   0.4950      0.0780   0.33700        0.04200   0.15000
#> 14  2020-05-01 13:00:00   0.5800      0.0870   0.33500        0.05000   0.16900
#> 15  2020-05-01 14:00:00   0.4710      0.0840   0.32700        0.03800   0.17000
#> 16  2020-05-01 15:00:00   0.4900      0.0840   0.42300        0.04100   0.29000
#> 17  2020-05-01 16:00:00   0.6610      0.0930   0.46700        0.04100   0.28400
#> 18  2020-05-01 17:00:00   0.6220      0.0870   0.39200        0.03000   0.27000
#> 19  2020-05-01 18:00:00   0.5130      0.0840   0.58100        0.03900   0.27000
#> 20  2020-05-01 19:00:00   1.0180      0.0920   0.79200        0.04500   0.39800
#> 21  2020-05-01 20:00:00   1.2310      0.0980   1.10400        0.04600   0.60300
#> 22  2020-05-01 21:00:00   1.1780      0.0950   1.02200        0.07300   0.57000
#> 23  2020-05-01 22:00:00   1.2670      0.0970   0.76500        0.05800   0.36400
#> 24  2020-05-01 23:00:00   0.9510      0.1030   0.65000        0.04200   0.57400
#> 25  2020-05-02 00:00:00   0.5830      0.1040   0.53400        0.03700   0.34200
#> 26  2020-05-02 01:00:00   0.4145      0.1375   0.68825        0.03450   0.25250
#> 27  2020-05-02 02:00:00   0.4990      0.1370   0.58300        0.03800   0.21100
#> 28  2020-05-02 03:00:00   0.6680      0.1360   0.37250        0.04500   0.12800
#> 29  2020-05-02 04:00:00   0.7530      0.1430   0.43900        0.04200   0.20400
#> 30  2020-05-02 05:00:00   0.9230      0.1570   0.57200        0.03600   0.35600
#> 31  2020-05-02 06:00:00   0.8910      0.1360   0.67300        0.06000   0.49700
#> 32  2020-05-02 07:00:00   1.1335      0.1180   0.54750        0.08150   0.96850
#> 33  2020-05-02 08:00:00   1.0770      0.1290   0.61500        0.06700   0.88000
#> 34  2020-05-02 09:00:00   0.9640      0.1510   0.75000        0.03800   0.70300
#> 35  2020-05-02 10:00:00   0.9720      0.1520   0.72700        0.04000   0.64300
#> 36  2020-05-02 11:00:00   1.3170      0.1670   0.83000        0.03800   0.58800
#> 37  2020-05-02 12:00:00   1.1750      0.1180   0.68700        0.04600   0.65600
#> 38  2020-05-02 13:00:00   0.5970      0.0900   0.41100        0.04000   0.35600
#> 39  2020-05-02 14:00:00   0.3920      0.0740   0.32500        0.04300   0.18700
#> 40  2020-05-02 15:00:00   0.3280      0.0730   0.28500        0.04500   0.17100
#> 41  2020-05-02 16:00:00   0.3410      0.0780   0.28600        0.03300   0.18800
#> 42  2020-05-02 17:00:00   0.6040      0.0750   0.34100        0.04400   0.22000
#> 43  2020-05-02 18:00:00   0.7120      0.0800   0.42000        0.04500   0.32200
#> 44  2020-05-02 19:00:00   1.0560      0.0880   0.63100        0.05600   0.29900
#> 45  2020-05-02 20:00:00   1.7950      0.0970   1.50800        0.06800   0.67500
#> 46  2020-05-02 21:00:00   1.6760      0.0940   1.56900        0.09800   0.52400
#> 47  2020-05-02 22:00:00   1.5840      0.1040   1.13400        0.05300   0.42000
#> 48  2020-05-02 23:00:00   2.8720      0.1155   1.55375        0.14425   0.34775
#> 49  2020-05-03 00:00:00   2.8230      0.1150   1.60850        0.13850   0.35550
#> 50  2020-05-03 01:00:00   2.7250      0.1140   1.71800        0.12700   0.37100
#> 51  2020-05-03 02:00:00   2.5290      0.1120   1.93700        0.10400   0.40200
#> 52  2020-05-03 03:00:00   5.6250      0.1220   4.06300        0.20100   0.48400
#> 53  2020-05-03 04:00:00   2.9010      0.1110   2.01200        0.10400   0.43700
#> 54  2020-05-03 05:00:00   2.5470      0.1210   2.23100        0.08500   0.56200
#> 55  2020-05-03 06:00:00   2.1540      0.1410   2.25900        0.14000   0.95000
#> 56  2020-05-03 07:00:00   1.4860      0.1110   1.16800        0.07500   0.47800
#> 57  2020-05-03 08:00:00   0.8320      0.0890   0.43700        0.03800   0.31000
#> 58  2020-05-03 09:00:00   0.9140      0.0840   0.45700        0.04700   0.37000
#> 59  2020-05-03 10:00:00   0.5360      0.0800   0.40500        0.03700   0.33000
#> 60  2020-05-03 11:00:00   0.5590      0.0780   0.30900        0.05000   0.30200
#> 61  2020-05-03 12:00:00   0.4970      0.0760   0.37300        0.05000   0.29000
#> 62  2020-05-03 13:00:00   0.4490      0.0750   0.27400        0.03700   0.32000
#> 63  2020-05-03 14:00:00   0.3920      0.0750   0.28400        0.04700   0.34000
#> 64  2020-05-03 15:00:00   0.4700      0.0660   0.29500        0.03500   0.30900
#> 65  2020-05-03 16:00:00   0.7550      0.0760   0.35700        0.04200   0.36000
#> 66  2020-05-03 17:00:00   0.5430      0.0790   0.40900        0.03800   0.44300
#> 67  2020-05-03 18:00:00   0.8500      0.0760   0.47100        0.03900   0.45800
#> 68  2020-05-03 19:00:00   0.6080      0.0770   0.52200        0.04700   0.49900
#> 69  2020-05-03 20:00:00   0.6770      0.0850   0.57200        0.03800   0.47300
#> 70  2020-05-03 21:00:00   0.5530      0.0770   0.49100        0.05600   0.43200
#> 71  2020-05-03 22:00:00   0.6090      0.0760   0.36400        0.03200   0.39900
#> 72  2020-05-03 23:00:00   0.5010      0.0730   0.34100        0.03800   0.39100
#> 73  2020-05-04 00:00:00   0.5130      0.0735   0.33800        0.05150   0.44500
#> 74  2020-05-04 01:00:00   0.5190      0.0750   0.32500        0.04800   0.43300
#> 75  2020-05-04 02:00:00   0.5310      0.0780   0.29900        0.04100   0.40900
#> 76  2020-05-04 03:00:00   0.5600      0.0810   0.33700        0.03800   0.45300
#> 77  2020-05-04 04:00:00   0.4490      0.0760   0.30400        0.03000   0.41800
#> 78  2020-05-04 05:00:00   0.6170      0.0840   0.35500        0.04300   0.69200
#> 79  2020-05-04 06:00:00   0.7020      0.0950   0.39900        0.05900   0.63800
#> 80  2020-05-04 07:00:00   0.8240      0.1110   0.45600        0.05100   0.49400
#> 81  2020-05-04 08:00:00   0.8810      0.1060   0.47800        0.03800   0.48700
#> 82  2020-05-04 09:00:00   1.0220      0.0940   0.51000        0.06700   0.48500
#> 83  2020-05-04 10:00:00   0.9470      0.0880   0.55200        0.04000   0.50600
#> 84  2020-05-04 11:00:00   0.6600      0.0800   0.40300        0.03200   0.54800
#> 85  2020-05-04 12:00:00   0.5820      0.0840   0.55600        0.05100   0.39300
#> 86  2020-05-04 13:00:00   0.5680      0.0790   0.38000        0.04300   0.42300
#> 87  2020-05-04 14:00:00   0.6350      0.0810   0.32200        0.03900   0.35700
#> 88  2020-05-04 15:00:00   0.6350      0.0810   0.34200        0.03800   0.38600
#> 89  2020-05-04 16:00:00   0.4810      0.0740   0.34600        0.03000   0.30900
#> 90  2020-05-04 17:00:00   0.4850      0.0750   0.34300        0.04400   0.24700
#> 91  2020-05-04 18:00:00   0.6180      0.0720   0.40500        0.03600   0.32800
#> 92  2020-05-04 19:00:00   0.7230      0.0740   0.55800        0.04000   0.31300
#> 93  2020-05-04 20:00:00   1.6200      0.0790   0.60900        0.06100   0.30900
#> 94  2020-05-04 21:00:00   0.8750      0.0770   0.48300        0.05000   0.36400
#> 95  2020-05-04 22:00:00   0.6530      0.0720   0.39500        0.04000   0.36300
#> 96  2020-05-04 23:00:00   1.1880      0.0830   0.54800        0.05400   0.34900
#> 97  2020-05-05 00:00:00   0.7430      0.0715   0.46100        0.05100   0.79850
#> 98  2020-05-05 01:00:00   0.9790      0.0860   0.58100        0.05000   0.73700
#> 99  2020-05-05 02:00:00   1.4510      0.1150   0.82100        0.04800   0.61400
#> 100 2020-05-05 03:00:00   3.1480      0.1540   1.52400        0.11900   0.55300
#> 101 2020-05-05 04:00:00   3.9000      0.2430   1.84600        0.08100   0.75400
#> 102 2020-05-05 05:00:00   4.4450      0.3530   1.97000        0.08900   0.97700
#> 103 2020-05-05 06:00:00   4.8240      0.3200   2.14700        0.09600   0.85000
#> 104 2020-05-05 07:00:00   1.9350      0.1230   0.86800        0.06200   0.65900
#> 105 2020-05-05 08:00:00   1.1750      0.0980   0.72300        0.04800   0.65500
#> 106 2020-05-05 09:00:00   0.9130      0.0980   0.49800        0.04200   0.50500
#> 107 2020-05-05 10:00:00   0.8920      0.1260   0.47900        0.07000   0.30500
#> 108 2020-05-05 11:00:00   0.4940      0.0820   0.41500        0.03100   0.24500
#> 109 2020-05-05 12:00:00   0.3980      0.0750   0.32800        0.09800   0.19500
#> 110 2020-05-05 13:00:00   0.4030      0.0780   0.27700        0.03000   0.14000
#> 111 2020-05-05 14:00:00   0.5390      0.0760   0.32800        0.04400   0.20300
#> 112 2020-05-05 15:00:00   0.6200      0.0760   0.39400        0.03600   0.21400
#> 113 2020-05-05 16:00:00   0.8450      0.0790   0.63300        0.05100   0.27500
#> 114 2020-05-05 17:00:00   1.1320      0.0850   0.51600        0.05700   0.25300
#> 115 2020-05-05 18:00:00   1.1560      0.0960   0.64800        0.04800   0.34800
#> 116 2020-05-05 19:00:00   0.9530      0.0970   0.73500        0.05800   0.41100
#> 117 2020-05-05 20:00:00   1.0180      0.1030   1.24100        0.04100   0.60700
#> 118 2020-05-05 21:00:00   1.3990      0.0930   0.99100        0.04900   0.37900
#> 119 2020-05-05 22:00:00   1.6050      0.0860   1.10700        0.05400   0.36200
#> 120 2020-05-05 23:00:00   2.0300      0.0980   1.40300        0.08200   0.40400
#> 
#> $Con_ppbv_stat
#>          species  Mean    sd   Min   Q25   Q50   Q75   Max Proportion
#> 1       n.Butane 1.081 0.903 0.328 0.560 0.806 1.175 5.625     0.4598
#> 3      Propylene 0.700 0.563 0.233 0.373 0.513 0.729 4.063     0.2977
#> 5      Acetylene 0.414 0.184 0.112 0.292 0.375 0.500 0.977     0.1761
#> 2    Cyclohexane 0.102 0.040 0.066 0.078 0.090 0.112 0.353     0.0434
#> 4 trans.2.Butene 0.054 0.027 0.020 0.038 0.045 0.056 0.201     0.0230
#> 
#> $Con_ppbv_group
#>                    Time Alkanes Alkenes_exclude_BVOC Alkynes
#> 1   2020-05-01 00:00:00  0.6460              0.25300 0.17500
#> 2   2020-05-01 01:00:00  0.8085              0.40400 0.20250
#> 3   2020-05-01 02:00:00  0.9710              0.55500 0.23000
#> 4   2020-05-01 03:00:00  1.2960              0.85700 0.28500
#> 5   2020-05-01 04:00:00  1.4270              0.73300 0.29200
#> 6   2020-05-01 05:00:00  1.0150              0.58700 0.37000
#> 7   2020-05-01 06:00:00  0.9660              0.57400 0.39700
#> 8   2020-05-01 07:00:00  0.9280              0.62600 0.47000
#> 9   2020-05-01 08:00:00  0.8690              0.65700 0.55100
#> 10  2020-05-01 09:00:00  1.0440              0.49900 0.51900
#> 11  2020-05-01 10:00:00  0.7660              0.36600 0.22300
#> 12  2020-05-01 11:00:00  0.6140              0.45100 0.11200
#> 13  2020-05-01 12:00:00  0.5730              0.37900 0.15000
#> 14  2020-05-01 13:00:00  0.6670              0.38500 0.16900
#> 15  2020-05-01 14:00:00  0.5550              0.36500 0.17000
#> 16  2020-05-01 15:00:00  0.5740              0.46400 0.29000
#> 17  2020-05-01 16:00:00  0.7540              0.50800 0.28400
#> 18  2020-05-01 17:00:00  0.7090              0.42200 0.27000
#> 19  2020-05-01 18:00:00  0.5970              0.62000 0.27000
#> 20  2020-05-01 19:00:00  1.1100              0.83700 0.39800
#> 21  2020-05-01 20:00:00  1.3290              1.15000 0.60300
#> 22  2020-05-01 21:00:00  1.2730              1.09500 0.57000
#> 23  2020-05-01 22:00:00  1.3640              0.82300 0.36400
#> 24  2020-05-01 23:00:00  1.0540              0.69200 0.57400
#> 25  2020-05-02 00:00:00  0.6870              0.57100 0.34200
#> 26  2020-05-02 01:00:00  0.5520              0.72275 0.25250
#> 27  2020-05-02 02:00:00  0.6360              0.62100 0.21100
#> 28  2020-05-02 03:00:00  0.8040              0.41750 0.12800
#> 29  2020-05-02 04:00:00  0.8960              0.48100 0.20400
#> 30  2020-05-02 05:00:00  1.0800              0.60800 0.35600
#> 31  2020-05-02 06:00:00  1.0270              0.73300 0.49700
#> 32  2020-05-02 07:00:00  1.2515              0.62900 0.96850
#> 33  2020-05-02 08:00:00  1.2060              0.68200 0.88000
#> 34  2020-05-02 09:00:00  1.1150              0.78800 0.70300
#> 35  2020-05-02 10:00:00  1.1240              0.76700 0.64300
#> 36  2020-05-02 11:00:00  1.4840              0.86800 0.58800
#> 37  2020-05-02 12:00:00  1.2930              0.73300 0.65600
#> 38  2020-05-02 13:00:00  0.6870              0.45100 0.35600
#> 39  2020-05-02 14:00:00  0.4660              0.36800 0.18700
#> 40  2020-05-02 15:00:00  0.4010              0.33000 0.17100
#> 41  2020-05-02 16:00:00  0.4190              0.31900 0.18800
#> 42  2020-05-02 17:00:00  0.6790              0.38500 0.22000
#> 43  2020-05-02 18:00:00  0.7920              0.46500 0.32200
#> 44  2020-05-02 19:00:00  1.1440              0.68700 0.29900
#> 45  2020-05-02 20:00:00  1.8920              1.57600 0.67500
#> 46  2020-05-02 21:00:00  1.7700              1.66700 0.52400
#> 47  2020-05-02 22:00:00  1.6880              1.18700 0.42000
#> 48  2020-05-02 23:00:00  2.9875              1.69800 0.34775
#> 49  2020-05-03 00:00:00  2.9380              1.74700 0.35550
#> 50  2020-05-03 01:00:00  2.8390              1.84500 0.37100
#> 51  2020-05-03 02:00:00  2.6410              2.04100 0.40200
#> 52  2020-05-03 03:00:00  5.7470              4.26400 0.48400
#> 53  2020-05-03 04:00:00  3.0120              2.11600 0.43700
#> 54  2020-05-03 05:00:00  2.6680              2.31600 0.56200
#> 55  2020-05-03 06:00:00  2.2950              2.39900 0.95000
#> 56  2020-05-03 07:00:00  1.5970              1.24300 0.47800
#> 57  2020-05-03 08:00:00  0.9210              0.47500 0.31000
#> 58  2020-05-03 09:00:00  0.9980              0.50400 0.37000
#> 59  2020-05-03 10:00:00  0.6160              0.44200 0.33000
#> 60  2020-05-03 11:00:00  0.6370              0.35900 0.30200
#> 61  2020-05-03 12:00:00  0.5730              0.42300 0.29000
#> 62  2020-05-03 13:00:00  0.5240              0.31100 0.32000
#> 63  2020-05-03 14:00:00  0.4670              0.33100 0.34000
#> 64  2020-05-03 15:00:00  0.5360              0.33000 0.30900
#> 65  2020-05-03 16:00:00  0.8310              0.39900 0.36000
#> 66  2020-05-03 17:00:00  0.6220              0.44700 0.44300
#> 67  2020-05-03 18:00:00  0.9260              0.51000 0.45800
#> 68  2020-05-03 19:00:00  0.6850              0.56900 0.49900
#> 69  2020-05-03 20:00:00  0.7620              0.61000 0.47300
#> 70  2020-05-03 21:00:00  0.6300              0.54700 0.43200
#> 71  2020-05-03 22:00:00  0.6850              0.39600 0.39900
#> 72  2020-05-03 23:00:00  0.5740              0.37900 0.39100
#> 73  2020-05-04 00:00:00  0.5865              0.38950 0.44500
#> 74  2020-05-04 01:00:00  0.5940              0.37300 0.43300
#> 75  2020-05-04 02:00:00  0.6090              0.34000 0.40900
#> 76  2020-05-04 03:00:00  0.6410              0.37500 0.45300
#> 77  2020-05-04 04:00:00  0.5250              0.33400 0.41800
#> 78  2020-05-04 05:00:00  0.7010              0.39800 0.69200
#> 79  2020-05-04 06:00:00  0.7970              0.45800 0.63800
#> 80  2020-05-04 07:00:00  0.9350              0.50700 0.49400
#> 81  2020-05-04 08:00:00  0.9870              0.51600 0.48700
#> 82  2020-05-04 09:00:00  1.1160              0.57700 0.48500
#> 83  2020-05-04 10:00:00  1.0350              0.59200 0.50600
#> 84  2020-05-04 11:00:00  0.7400              0.43500 0.54800
#> 85  2020-05-04 12:00:00  0.6660              0.60700 0.39300
#> 86  2020-05-04 13:00:00  0.6470              0.42300 0.42300
#> 87  2020-05-04 14:00:00  0.7160              0.36100 0.35700
#> 88  2020-05-04 15:00:00  0.7160              0.38000 0.38600
#> 89  2020-05-04 16:00:00  0.5550              0.37600 0.30900
#> 90  2020-05-04 17:00:00  0.5600              0.38700 0.24700
#> 91  2020-05-04 18:00:00  0.6900              0.44100 0.32800
#> 92  2020-05-04 19:00:00  0.7970              0.59800 0.31300
#> 93  2020-05-04 20:00:00  1.6990              0.67000 0.30900
#> 94  2020-05-04 21:00:00  0.9520              0.53300 0.36400
#> 95  2020-05-04 22:00:00  0.7250              0.43500 0.36300
#> 96  2020-05-04 23:00:00  1.2710              0.60200 0.34900
#> 97  2020-05-05 00:00:00  0.8145              0.51200 0.79850
#> 98  2020-05-05 01:00:00  1.0650              0.63100 0.73700
#> 99  2020-05-05 02:00:00  1.5660              0.86900 0.61400
#> 100 2020-05-05 03:00:00  3.3020              1.64300 0.55300
#> 101 2020-05-05 04:00:00  4.1430              1.92700 0.75400
#> 102 2020-05-05 05:00:00  4.7980              2.05900 0.97700
#> 103 2020-05-05 06:00:00  5.1440              2.24300 0.85000
#> 104 2020-05-05 07:00:00  2.0580              0.93000 0.65900
#> 105 2020-05-05 08:00:00  1.2730              0.77100 0.65500
#> 106 2020-05-05 09:00:00  1.0110              0.54000 0.50500
#> 107 2020-05-05 10:00:00  1.0180              0.54900 0.30500
#> 108 2020-05-05 11:00:00  0.5760              0.44600 0.24500
#> 109 2020-05-05 12:00:00  0.4730              0.42600 0.19500
#> 110 2020-05-05 13:00:00  0.4810              0.30700 0.14000
#> 111 2020-05-05 14:00:00  0.6150              0.37200 0.20300
#> 112 2020-05-05 15:00:00  0.6960              0.43000 0.21400
#> 113 2020-05-05 16:00:00  0.9240              0.68400 0.27500
#> 114 2020-05-05 17:00:00  1.2170              0.57300 0.25300
#> 115 2020-05-05 18:00:00  1.2520              0.69600 0.34800
#> 116 2020-05-05 19:00:00  1.0500              0.79300 0.41100
#> 117 2020-05-05 20:00:00  1.1210              1.28200 0.60700
#> 118 2020-05-05 21:00:00  1.4920              1.04000 0.37900
#> 119 2020-05-05 22:00:00  1.6910              1.16100 0.36200
#> 120 2020-05-05 23:00:00  2.1280              1.48500 0.40400
#> 
#> $Con_ppbv_group_stat
#>                Species  Mean    SD   Min   Q25   Q50   Q75   Max Proportion
#> 1              Alkanes 1.182 0.931 0.401 0.645 0.922 1.271 5.747     0.5030
#> 2 Alkenes_exclude_BVOC 0.754 0.587 0.253 0.421 0.562 0.775 4.264     0.3209
#> 3              Alkynes 0.414 0.184 0.112 0.292 0.375 0.500 0.977     0.1762
#>