Calculate OH reactivity

Calculate OH reactivity of VOC time series in 25 degree celsius. Note: for Chinese VOC name, please also use English punctuation.

Usage

loh(
  df,
  unit = "ppbv",
  t = 25,
  p = 101.325,
  stcd = FALSE,
  sortd = TRUE,
  atk = TRUE,
  chn = FALSE,
  bvoc = TRUE
)

Arguments

df

dataframe contains time series.

unit

unit for VOC concentration. A character vector from these options: "ugm" or "ppbv". "ugm" means ug/m3. "ppbv" means part per billion volumn.

t

Temperature, in Degrees Celsius, used to convert data in micrograms per cubic meter to standard conditions (25 Degrees Celsius, 101.325 kPa). By default, t equals to 25 Degrees Celsius.

p

Pressure, in kPa, used to convert data in micrograms per cubic meter to standard conditions (25 Degrees Celsius, 101.325 kPa). By default, p equals to 101.325 kPa.

stcd

logical. Does it output the concentration in standard condition? The default vaule is FALSE.

sortd

logical value. It determines whether the VOC species are sorted or not. By default, sortd has value "TRUE". If TRUE, VOC species in time series will be arranged according to VOC group, relative molecular weight, and OH Rate Constant.

atk

logical. use kOH value from atk or not? If not, kOH comes from 'AopWin v1.92' will be used. The default vaule is TRUE.

chn

logical. Dose colnames present as Chinese? The default vaule is FALSE.

bvoc

logical. Whether you want to list BVOC as a separate VOC group? The default vaule is TRUE.

Value

a list contains 5 tables: KOH_Result: the matched KOH value results; LOH_Result: the LOH time series of VOC by species; LOH_Result_stat: the statistics of LOH of VOC by species; LOH_Result_group: the LOH time series of VOC classified by groups; LOH_Result_group_mean: the statistics of LOH of VOC according to major groups.

Details

The CAS number is matched for each VOC speices (from column name), and the OH Rate Constant is matched through the CAS number and used for time series calculation.
The OH Rate Constant comes from 'AopWin v1.92' in 25 degree celsius.

Examples

loh(voc)
#> $KOH_Result
#>             Name      CAS   Matched_Name      Koh Koh_type    MW   Group
#> 3       n-Butane 106-97-8       n-butane 2.36e-12 Atkinson 58.12 Alkanes
#> 5    Cyclohexane 110-82-7    cyclohexane 6.97e-12 Atkinson 84.16 Alkanes
#> 1      Propylene 115-07-1        propene 2.63e-11 Atkinson 42.08 Alkenes
#> 4 trans-2-Butene 624-64-6 trans-2-butene 6.40e-11 Atkinson 56.11 Alkenes
#> 2      Acetylene  74-86-2      acetylene 8.15e-13   AopWin 26.04 Alkynes
#>   Raw_order
#> 3         3
#> 5         5
#> 1         1
#> 4         4
#> 2         2
#> 
#> $LOH_Result
#>                    Time   n.Butane Cyclohexane Propylene trans.2.Butene
#> 1   2020-05-01 00:00:00 0.03162059  0.01751024 0.1509283     0.03152601
#> 2   2020-05-01 01:00:00 0.04092076  0.01793941 0.2435581     0.04413641
#> 3   2020-05-01 02:00:00 0.05022093  0.01836858 0.3361880     0.05674682
#> 4   2020-05-01 03:00:00 0.06882128  0.01922693 0.5214476     0.08196763
#> 5   2020-05-01 04:00:00 0.07579641  0.02111528 0.4262267     0.11822254
#> 6   2020-05-01 05:00:00 0.05254598  0.01905526 0.3484954     0.07723873
#> 7   2020-05-01 06:00:00 0.04870966  0.02197363 0.3452566     0.06462832
#> 8   2020-05-01 07:00:00 0.04580335  0.02403366 0.3718148     0.08196763
#> 9   2020-05-01 08:00:00 0.04359456  0.02042861 0.4035551     0.05359422
#> 10  2020-05-01 09:00:00 0.05498727  0.01682356 0.2895492     0.08196763
#> 11  2020-05-01 10:00:00 0.03975824  0.01407686 0.2137611     0.05674682
#> 12  2020-05-01 11:00:00 0.03103933  0.01373352 0.2642865     0.06778092
#> 13  2020-05-01 12:00:00 0.02877241  0.01339018 0.2182955     0.06620462
#> 14  2020-05-01 13:00:00 0.03371313  0.01493520 0.2169999     0.07881503
#> 15  2020-05-01 14:00:00 0.02737738  0.01442019 0.2118178     0.05989942
#> 16  2020-05-01 15:00:00 0.02848178  0.01442019 0.2740029     0.06462832
#> 17  2020-05-01 16:00:00 0.03842134  0.01596522 0.3025044     0.06462832
#> 18  2020-05-01 17:00:00 0.03615442  0.01493520 0.2539223     0.04728902
#> 19  2020-05-01 18:00:00 0.02981868  0.01442019 0.3763491     0.06147572
#> 20  2020-05-01 19:00:00 0.05917235  0.01579355 0.5130267     0.07093352
#> 21  2020-05-01 20:00:00 0.07155321  0.01682356 0.7151281     0.07250982
#> 22  2020-05-01 21:00:00 0.06847252  0.01630855 0.6620117     0.11506994
#> 23  2020-05-01 22:00:00 0.07364575  0.01665189 0.4955372     0.09142543
#> 24  2020-05-01 23:00:00 0.05527790  0.01768190 0.4210446     0.06620462
#> 25  2020-05-02 00:00:00 0.03388751  0.01785357 0.3459044     0.05832312
#> 26  2020-05-02 01:00:00 0.02409326  0.02360448 0.4458215     0.05438237
#> 27  2020-05-02 02:00:00 0.02900491  0.02351865 0.3776447     0.05989942
#> 28  2020-05-02 03:00:00 0.03882822  0.02334698 0.2412910     0.07093352
#> 29  2020-05-02 04:00:00 0.04376894  0.02454866 0.2843671     0.06620462
#> 30  2020-05-02 05:00:00 0.05365037  0.02695203 0.3705193     0.05674682
#> 31  2020-05-02 06:00:00 0.05179034  0.02334698 0.4359432     0.09457803
#> 32  2020-05-02 07:00:00 0.06588591  0.02025694 0.3546491     0.12846849
#> 33  2020-05-02 08:00:00 0.06260179  0.02214530 0.3983730     0.10561214
#> 34  2020-05-02 09:00:00 0.05603354  0.02592202 0.4858207     0.05989942
#> 35  2020-05-02 10:00:00 0.05649855  0.02609368 0.4709222     0.06305202
#> 36  2020-05-02 11:00:00 0.07655205  0.02866872 0.5376416     0.05989942
#> 37  2020-05-02 12:00:00 0.06829815  0.02025694 0.4450118     0.07250982
#> 38  2020-05-02 13:00:00 0.03470127  0.01545021 0.2662298     0.06305202
#> 39  2020-05-02 14:00:00 0.02278542  0.01270350 0.2105223     0.06778092
#> 40  2020-05-02 15:00:00 0.01906535  0.01253184 0.1846119     0.07093352
#> 41  2020-05-02 16:00:00 0.01982099  0.01339018 0.1852596     0.05201792
#> 42  2020-05-02 17:00:00 0.03510815  0.01287517 0.2208865     0.06935722
#> 43  2020-05-02 18:00:00 0.04138577  0.01373352 0.2720596     0.07093352
#> 44  2020-05-02 19:00:00 0.06138114  0.01510687 0.4087372     0.08827283
#> 45  2020-05-02 20:00:00 0.10433632  0.01665189 0.9768236     0.10718844
#> 46  2020-05-02 21:00:00 0.09741931  0.01613688 1.0163370     0.15447745
#> 47  2020-05-02 22:00:00 0.09207171  0.01785357 0.7345610     0.08354393
#> 48  2020-05-02 23:00:00 0.16693811  0.01982777 1.0064586     0.22738135
#> 49  2020-05-03 00:00:00 0.16408993  0.01974193 1.0419236     0.21831762
#> 50  2020-05-03 01:00:00 0.15839357  0.01957026 1.1128534     0.20019017
#> 51  2020-05-03 02:00:00 0.14700086  0.01922693 1.2547130     0.16393525
#> 52  2020-05-03 03:00:00 0.32695921  0.02094362 2.6318529     0.31683641
#> 53  2020-05-03 04:00:00 0.16862376  0.01905526 1.3032951     0.16393525
#> 54  2020-05-03 05:00:00 0.14804713  0.02077195 1.4451548     0.13398554
#> 55  2020-05-03 06:00:00 0.12520358  0.02420533 1.4632921     0.22068207
#> 56  2020-05-03 07:00:00 0.08637536  0.01905526 0.7565848     0.11822254
#> 57  2020-05-03 08:00:00 0.04836090  0.01527854 0.2830716     0.05989942
#> 58  2020-05-03 09:00:00 0.05312724  0.01442019 0.2960268     0.07408612
#> 59  2020-05-03 10:00:00 0.03115558  0.01373352 0.2623432     0.05832312
#> 60  2020-05-03 11:00:00 0.03249248  0.01339018 0.2001581     0.07881503
#> 61  2020-05-03 12:00:00 0.02888866  0.01304684 0.2416149     0.07881503
#> 62  2020-05-03 13:00:00 0.02609861  0.01287517 0.1774865     0.05832312
#> 63  2020-05-03 14:00:00 0.02278542  0.01287517 0.1839641     0.07408612
#> 64  2020-05-03 15:00:00 0.02731926  0.01133015 0.1910895     0.05517052
#> 65  2020-05-03 16:00:00 0.04388519  0.01304684 0.2312507     0.06620462
#> 66  2020-05-03 17:00:00 0.03156246  0.01356185 0.2649342     0.05989942
#> 67  2020-05-03 18:00:00 0.04940717  0.01304684 0.3050954     0.06147572
#> 68  2020-05-03 19:00:00 0.03534066  0.01321851 0.3381312     0.07408612
#> 69  2020-05-03 20:00:00 0.03935136  0.01459186 0.3705193     0.05989942
#> 70  2020-05-03 21:00:00 0.03214372  0.01321851 0.3180506     0.08827283
#> 71  2020-05-03 22:00:00 0.03539878  0.01304684 0.2357850     0.05044162
#> 72  2020-05-03 23:00:00 0.02912117  0.01253184 0.2208865     0.05989942
#> 73  2020-05-04 00:00:00 0.02981868  0.01261767 0.2189432     0.08117948
#> 74  2020-05-04 01:00:00 0.03016744  0.01287517 0.2105223     0.07566243
#> 75  2020-05-04 02:00:00 0.03086495  0.01339018 0.1936805     0.06462832
#> 76  2020-05-04 03:00:00 0.03255061  0.01390519 0.2182955     0.05989942
#> 77  2020-05-04 04:00:00 0.02609861  0.01304684 0.1969193     0.04728902
#> 78  2020-05-04 05:00:00 0.03586379  0.01442019 0.2299552     0.06778092
#> 79  2020-05-04 06:00:00 0.04080451  0.01630855 0.2584566     0.09300173
#> 80  2020-05-04 07:00:00 0.04789589  0.01905526 0.2953790     0.08039133
#> 81  2020-05-04 08:00:00 0.05120908  0.01819691 0.3096298     0.05989942
#> 82  2020-05-04 09:00:00 0.05940486  0.01613688 0.3303581     0.10561214
#> 83  2020-05-04 10:00:00 0.05504540  0.01510687 0.3575641     0.06305202
#> 84  2020-05-04 11:00:00 0.03836321  0.01373352 0.2610477     0.05044162
#> 85  2020-05-04 12:00:00 0.03382938  0.01442019 0.3601551     0.08039133
#> 86  2020-05-04 13:00:00 0.03301561  0.01356185 0.2461492     0.06778092
#> 87  2020-05-04 14:00:00 0.03691006  0.01390519 0.2085790     0.06147572
#> 88  2020-05-04 15:00:00 0.03691006  0.01390519 0.2215343     0.05989942
#> 89  2020-05-04 16:00:00 0.02795865  0.01270350 0.2241253     0.04728902
#> 90  2020-05-04 17:00:00 0.02819115  0.01287517 0.2221820     0.06935722
#> 91  2020-05-04 18:00:00 0.03592192  0.01236017 0.2623432     0.05674682
#> 92  2020-05-04 19:00:00 0.04202516  0.01270350 0.3614506     0.06305202
#> 93  2020-05-04 20:00:00 0.09416425  0.01356185 0.3944864     0.09615433
#> 94  2020-05-04 21:00:00 0.05086032  0.01321851 0.3128686     0.07881503
#> 95  2020-05-04 22:00:00 0.03795633  0.01236017 0.2558656     0.06305202
#> 96  2020-05-04 23:00:00 0.06905379  0.01424853 0.3549730     0.08512023
#> 97  2020-05-05 00:00:00 0.04318768  0.01227433 0.2986178     0.08039133
#> 98  2020-05-05 01:00:00 0.05690543  0.01476353 0.3763491     0.07881503
#> 99  2020-05-05 02:00:00 0.08434094  0.01974193 0.5318118     0.07566243
#> 100 2020-05-05 03:00:00 0.18298091  0.02643702 0.9871878     0.18757976
#> 101 2020-05-05 04:00:00 0.22669172  0.04171556 1.1957668     0.12768034
#> 102 2020-05-05 05:00:00 0.25837043  0.06059915 1.2760892     0.14029075
#> 103 2020-05-05 06:00:00 0.28040022  0.05493407 1.3907429     0.15132485
#> 104 2020-05-05 07:00:00 0.11247397  0.02111528 0.5622565     0.09773063
#> 105 2020-05-05 08:00:00 0.06829815  0.01682356 0.4683312     0.07566243
#> 106 2020-05-05 09:00:00 0.05306911  0.01682356 0.3225850     0.06620462
#> 107 2020-05-05 10:00:00 0.05184846  0.02163029 0.3102775     0.11034104
#> 108 2020-05-05 11:00:00 0.02871428  0.01407686 0.2688208     0.04886532
#> 109 2020-05-05 12:00:00 0.02313418  0.01287517 0.2124656     0.15447745
#> 110 2020-05-05 13:00:00 0.02342481  0.01339018 0.1794298     0.04728902
#> 111 2020-05-05 14:00:00 0.03132996  0.01304684 0.2124656     0.06935722
#> 112 2020-05-05 15:00:00 0.03603817  0.01304684 0.2552178     0.05674682
#> 113 2020-05-05 16:00:00 0.04911654  0.01356185 0.4100327     0.08039133
#> 114 2020-05-05 17:00:00 0.06579872  0.01459186 0.3342447     0.08984913
#> 115 2020-05-05 18:00:00 0.06719375  0.01648022 0.4197491     0.07566243
#> 116 2020-05-05 19:00:00 0.05539416  0.01665189 0.4761043     0.09142543
#> 117 2020-05-05 20:00:00 0.05917235  0.01768190 0.8038714     0.06462832
#> 118 2020-05-05 21:00:00 0.08131839  0.01596522 0.6419311     0.07723873
#> 119 2020-05-05 22:00:00 0.09329236  0.01476353 0.7170714     0.08512023
#> 120 2020-05-05 23:00:00 0.11799595  0.01682356 0.9088087     0.12925664
#>       Acetylene
#> 1   0.003512810
#> 2   0.004064823
#> 3   0.004616836
#> 4   0.005720863
#> 5   0.005861375
#> 6   0.007427085
#> 7   0.007969061
#> 8   0.009434405
#> 9   0.011060334
#> 10  0.010417992
#> 11  0.004476324
#> 12  0.002248199
#> 13  0.003010980
#> 14  0.003392371
#> 15  0.003412444
#> 16  0.005821229
#> 17  0.005700789
#> 18  0.005419765
#> 19  0.005419765
#> 20  0.007989134
#> 21  0.012104141
#> 22  0.011441725
#> 23  0.007306646
#> 24  0.011522018
#> 25  0.006865035
#> 26  0.005068483
#> 27  0.004235446
#> 28  0.002569370
#> 29  0.004094933
#> 30  0.007146060
#> 31  0.009976381
#> 32  0.019440896
#> 33  0.017664418
#> 34  0.014111461
#> 35  0.012907069
#> 36  0.011803043
#> 37  0.013168020
#> 38  0.007146060
#> 39  0.003753689
#> 40  0.003432518
#> 41  0.003773762
#> 42  0.004416104
#> 43  0.006463571
#> 44  0.006001887
#> 45  0.013549411
#> 46  0.010518358
#> 47  0.008430745
#> 48  0.006980456
#> 49  0.007136023
#> 50  0.007447158
#> 51  0.008069427
#> 52  0.009715430
#> 53  0.008771989
#> 54  0.011281140
#> 55  0.019069542
#> 56  0.009594991
#> 57  0.006222693
#> 58  0.007427085
#> 59  0.006624157
#> 60  0.006062107
#> 61  0.005821229
#> 62  0.006423425
#> 63  0.006824889
#> 64  0.006202619
#> 65  0.007226353
#> 66  0.008892428
#> 67  0.009193527
#> 68  0.010016528
#> 69  0.009494625
#> 70  0.008671623
#> 71  0.008009208
#> 72  0.007848622
#> 73  0.008932575
#> 74  0.008691696
#> 75  0.008209940
#> 76  0.009093160
#> 77  0.008390598
#> 78  0.013890656
#> 79  0.012806703
#> 80  0.009916162
#> 81  0.009775649
#> 82  0.009735503
#> 83  0.010157040
#> 84  0.011000115
#> 85  0.007888768
#> 86  0.008490964
#> 87  0.007166133
#> 88  0.007748256
#> 89  0.006202619
#> 90  0.004958081
#> 91  0.006584010
#> 92  0.006282912
#> 93  0.006202619
#> 94  0.007306646
#> 95  0.007286572
#> 96  0.007005547
#> 97  0.016028452
#> 98  0.014793950
#> 99  0.012324946
#> 100 0.011100481
#> 101 0.015135194
#> 102 0.019611518
#> 103 0.017062222
#> 104 0.013228240
#> 105 0.013147947
#> 106 0.010136967
#> 107 0.006122327
#> 108 0.004917934
#> 109 0.003914274
#> 110 0.002810248
#> 111 0.004074860
#> 112 0.004295665
#> 113 0.005520131
#> 114 0.005078520
#> 115 0.006985474
#> 116 0.008250086
#> 117 0.012184434
#> 118 0.007607744
#> 119 0.007266499
#> 120 0.008109574
#> 
#> $LOH_Result_stat
#>          Species  Mean    sd   Min   Q25   Q50   Q75   Max Proportion
#> 3      Propylene 0.454 0.365 0.151 0.242 0.332 0.472 2.632     0.7241
#> 4 trans.2.Butene 0.085 0.043 0.032 0.060 0.071 0.089 0.317     0.1356
#> 1       n.Butane 0.063 0.053 0.019 0.033 0.047 0.068 0.327     0.1005
#> 2    Cyclohexane 0.017 0.007 0.011 0.013 0.015 0.019 0.061     0.0271
#> 5      Acetylene 0.008 0.004 0.002 0.006 0.008 0.010 0.020     0.0128
#> 
#> $LOH_Result_group
#>                    Time    Alkanes Alkenes_exclude_BVOC     Alkynes
#> 1   2020-05-01 00:00:00 0.04913082            0.1824543 0.003512810
#> 2   2020-05-01 01:00:00 0.05886017            0.2876945 0.004064823
#> 3   2020-05-01 02:00:00 0.06858952            0.3929348 0.004616836
#> 4   2020-05-01 03:00:00 0.08804821            0.6034152 0.005720863
#> 5   2020-05-01 04:00:00 0.09691170            0.5444493 0.005861375
#> 6   2020-05-01 05:00:00 0.07160123            0.4257341 0.007427085
#> 7   2020-05-01 06:00:00 0.07068329            0.4098849 0.007969061
#> 8   2020-05-01 07:00:00 0.06983701            0.4537824 0.009434405
#> 9   2020-05-01 08:00:00 0.06402317            0.4571493 0.011060334
#> 10  2020-05-01 09:00:00 0.07181083            0.3715168 0.010417992
#> 11  2020-05-01 10:00:00 0.05383510            0.2705079 0.004476324
#> 12  2020-05-01 11:00:00 0.04477285            0.3320674 0.002248199
#> 13  2020-05-01 12:00:00 0.04216259            0.2845001 0.003010980
#> 14  2020-05-01 13:00:00 0.04864833            0.2958150 0.003392371
#> 15  2020-05-01 14:00:00 0.04179758            0.2717173 0.003412444
#> 16  2020-05-01 15:00:00 0.04290197            0.3386312 0.005821229
#> 17  2020-05-01 16:00:00 0.05438656            0.3671327 0.005700789
#> 18  2020-05-01 17:00:00 0.05108962            0.3012113 0.005419765
#> 19  2020-05-01 18:00:00 0.04423887            0.4378249 0.005419765
#> 20  2020-05-01 19:00:00 0.07496590            0.5839602 0.007989134
#> 21  2020-05-01 20:00:00 0.08837677            0.7876380 0.012104141
#> 22  2020-05-01 21:00:00 0.08478108            0.7770817 0.011441725
#> 23  2020-05-01 22:00:00 0.09029764            0.5869626 0.007306646
#> 24  2020-05-01 23:00:00 0.07295981            0.4872493 0.011522018
#> 25  2020-05-02 00:00:00 0.05174108            0.4042275 0.006865035
#> 26  2020-05-02 01:00:00 0.04769775            0.5002039 0.005068483
#> 27  2020-05-02 02:00:00 0.05252356            0.4375441 0.004235446
#> 28  2020-05-02 03:00:00 0.06217520            0.3122245 0.002569370
#> 29  2020-05-02 04:00:00 0.06831760            0.3505717 0.004094933
#> 30  2020-05-02 05:00:00 0.08060240            0.4272661 0.007146060
#> 31  2020-05-02 06:00:00 0.07513732            0.5305212 0.009976381
#> 32  2020-05-02 07:00:00 0.08614285            0.4831176 0.019440896
#> 33  2020-05-02 08:00:00 0.08474709            0.5039851 0.017664418
#> 34  2020-05-02 09:00:00 0.08195556            0.5457202 0.014111461
#> 35  2020-05-02 10:00:00 0.08259224            0.5339743 0.012907069
#> 36  2020-05-02 11:00:00 0.10522077            0.5975410 0.011803043
#> 37  2020-05-02 12:00:00 0.08855509            0.5175216 0.013168020
#> 38  2020-05-02 13:00:00 0.05015148            0.3292818 0.007146060
#> 39  2020-05-02 14:00:00 0.03548893            0.2783032 0.003753689
#> 40  2020-05-02 15:00:00 0.03159719            0.2555454 0.003432518
#> 41  2020-05-02 16:00:00 0.03321117            0.2372776 0.003773762
#> 42  2020-05-02 17:00:00 0.04798333            0.2902437 0.004416104
#> 43  2020-05-02 18:00:00 0.05511929            0.3429931 0.006463571
#> 44  2020-05-02 19:00:00 0.07648801            0.4970100 0.006001887
#> 45  2020-05-02 20:00:00 0.12098821            1.0840120 0.013549411
#> 46  2020-05-02 21:00:00 0.11355620            1.1708145 0.010518358
#> 47  2020-05-02 22:00:00 0.10992529            0.8181049 0.008430745
#> 48  2020-05-02 23:00:00 0.18676587            1.2338400 0.006980456
#> 49  2020-05-03 00:00:00 0.18383186            1.2602412 0.007136023
#> 50  2020-05-03 01:00:00 0.17796384            1.3130436 0.007447158
#> 51  2020-05-03 02:00:00 0.16622779            1.4186483 0.008069427
#> 52  2020-05-03 03:00:00 0.34790283            2.9486893 0.009715430
#> 53  2020-05-03 04:00:00 0.18767902            1.4672304 0.008771989
#> 54  2020-05-03 05:00:00 0.16881908            1.5791403 0.011281140
#> 55  2020-05-03 06:00:00 0.14940891            1.6839742 0.019069542
#> 56  2020-05-03 07:00:00 0.10543061            0.8748074 0.009594991
#> 57  2020-05-03 08:00:00 0.06363944            0.3429710 0.006222693
#> 58  2020-05-03 09:00:00 0.06754743            0.3701129 0.007427085
#> 59  2020-05-03 10:00:00 0.04488910            0.3206663 0.006624157
#> 60  2020-05-03 11:00:00 0.04588266            0.2789732 0.006062107
#> 61  2020-05-03 12:00:00 0.04193551            0.3204299 0.005821229
#> 62  2020-05-03 13:00:00 0.03897378            0.2358096 0.006423425
#> 63  2020-05-03 14:00:00 0.03566060            0.2580502 0.006824889
#> 64  2020-05-03 15:00:00 0.03864941            0.2462600 0.006202619
#> 65  2020-05-03 16:00:00 0.05693203            0.2974553 0.007226353
#> 66  2020-05-03 17:00:00 0.04512431            0.3248337 0.008892428
#> 67  2020-05-03 18:00:00 0.06245401            0.3665711 0.009193527
#> 68  2020-05-03 19:00:00 0.04855917            0.4122174 0.010016528
#> 69  2020-05-03 20:00:00 0.05394322            0.4304187 0.009494625
#> 70  2020-05-03 21:00:00 0.04536223            0.4063235 0.008671623
#> 71  2020-05-03 22:00:00 0.04844563            0.2862266 0.008009208
#> 72  2020-05-03 23:00:00 0.04165300            0.2807859 0.007848622
#> 73  2020-05-04 00:00:00 0.04243635            0.3001227 0.008932575
#> 74  2020-05-04 01:00:00 0.04304261            0.2861847 0.008691696
#> 75  2020-05-04 02:00:00 0.04425513            0.2583089 0.008209940
#> 76  2020-05-04 03:00:00 0.04645579            0.2781949 0.009093160
#> 77  2020-05-04 04:00:00 0.03914545            0.2442084 0.008390598
#> 78  2020-05-04 05:00:00 0.05028399            0.2977361 0.013890656
#> 79  2020-05-04 06:00:00 0.05711306            0.3514584 0.012806703
#> 80  2020-05-04 07:00:00 0.06695115            0.3757703 0.009916162
#> 81  2020-05-04 08:00:00 0.06940599            0.3695292 0.009775649
#> 82  2020-05-04 09:00:00 0.07554174            0.4359702 0.009735503
#> 83  2020-05-04 10:00:00 0.07015227            0.4206161 0.010157040
#> 84  2020-05-04 11:00:00 0.05209673            0.3114893 0.011000115
#> 85  2020-05-04 12:00:00 0.04824957            0.4405464 0.007888768
#> 86  2020-05-04 13:00:00 0.04657746            0.3139301 0.008490964
#> 87  2020-05-04 14:00:00 0.05081525            0.2700548 0.007166133
#> 88  2020-05-04 15:00:00 0.05081525            0.2814337 0.007748256
#> 89  2020-05-04 16:00:00 0.04066215            0.2714143 0.006202619
#> 90  2020-05-04 17:00:00 0.04106632            0.2915392 0.004958081
#> 91  2020-05-04 18:00:00 0.04828209            0.3190900 0.006584010
#> 92  2020-05-04 19:00:00 0.05472866            0.4245027 0.006282912
#> 93  2020-05-04 20:00:00 0.10772610            0.4906408 0.006202619
#> 94  2020-05-04 21:00:00 0.06407883            0.3916836 0.007306646
#> 95  2020-05-04 22:00:00 0.05031650            0.3189176 0.007286572
#> 96  2020-05-04 23:00:00 0.08330231            0.4400933 0.007005547
#> 97  2020-05-05 00:00:00 0.05546201            0.3790091 0.016028452
#> 98  2020-05-05 01:00:00 0.07166897            0.4551642 0.014793950
#> 99  2020-05-05 02:00:00 0.10408288            0.6074742 0.012324946
#> 100 2020-05-05 03:00:00 0.20941793            1.1747675 0.011100481
#> 101 2020-05-05 04:00:00 0.26840728            1.3234471 0.015135194
#> 102 2020-05-05 05:00:00 0.31896958            1.4163799 0.019611518
#> 103 2020-05-05 06:00:00 0.33533429            1.5420677 0.017062222
#> 104 2020-05-05 07:00:00 0.13358925            0.6599872 0.013228240
#> 105 2020-05-05 08:00:00 0.08512171            0.5439936 0.013147947
#> 106 2020-05-05 09:00:00 0.06989267            0.3887896 0.010136967
#> 107 2020-05-05 10:00:00 0.07347876            0.4206186 0.006122327
#> 108 2020-05-05 11:00:00 0.04279114            0.3176861 0.004917934
#> 109 2020-05-05 12:00:00 0.03600935            0.3669431 0.003914274
#> 110 2020-05-05 13:00:00 0.03681499            0.2267188 0.002810248
#> 111 2020-05-05 14:00:00 0.04437680            0.2818228 0.004074860
#> 112 2020-05-05 15:00:00 0.04908501            0.3119647 0.004295665
#> 113 2020-05-05 16:00:00 0.06267839            0.4904240 0.005520131
#> 114 2020-05-05 17:00:00 0.08039059            0.4240938 0.005078520
#> 115 2020-05-05 18:00:00 0.08367397            0.4954116 0.006985474
#> 116 2020-05-05 19:00:00 0.07204605            0.5675298 0.008250086
#> 117 2020-05-05 20:00:00 0.07685426            0.8684997 0.012184434
#> 118 2020-05-05 21:00:00 0.09728360            0.7191699 0.007607744
#> 119 2020-05-05 22:00:00 0.10805589            0.8021917 0.007266499
#> 120 2020-05-05 23:00:00 0.13481951            1.0380653 0.008109574
#> 
#> $LOH_Result_group_stat
#>                Species  Mean    SD   Min   Q25   Q50   Q75   Max Proportion
#> 2 Alkenes_exclude_BVOC 0.539 0.402 0.182 0.309 0.408 0.545 2.949     0.8596
#> 1              Alkanes 0.080 0.057 0.032 0.048 0.064 0.085 0.348     0.1276
#> 3              Alkynes 0.008 0.004 0.002 0.006 0.008 0.010 0.020     0.0128
#>